% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{add_}
\alias{add_}
\alias{add_data}
\alias{add_metadata}
\alias{add_column_mappings}
\title{Add information to a \code{dwca} object}
\usage{
add_data(.dwca, df, file)

add_metadata(.dwca, xml, file, ...)

add_column_mappings(.dwca, xml, file, ...)
}
\arguments{
\item{.dwca}{An object of class \code{dwca}, constructed with \code{dwca()}.}

\item{df}{Optional R object; \code{tibble} or \code{data.frame}.}

\item{file}{Optional file name to be imported with \code{readr::read_csv()} for
\code{.csv}, \code{xml2::read_xml()} for \code{.xml}, or \code{galaxias::read_md()} for
\code{.md}, \code{.rmd} or \code{.qmd}.}

\item{xml}{Optional R object imported with \code{xml2} (class \code{xml_document}).}

\item{...}{Additional information passed to \code{read_csv()} or \code{read_xml()}.}
}
\description{
These are all pipable functions for adding relevant information to a \code{dwca}
object
}
