% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_random_identifier}
\alias{build_random_identifier}
\alias{build_composite_identifier}
\title{Functions to build new data for compatability with DwC}
\usage{
build_random_identifier(data, random_seed = NULL)

build_composite_identifier(data, .cols = NULL)
}
\arguments{
\item{data}{A tibble}

\item{random_seed}{Integer: what random seed should be used? Optional, but
useful for reproducibility.}

\item{.cols}{If \code{type = "merge"}, what columns should be used?}
}
\description{
These are risky. Use with care
}
