% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build a Darwin Core Archive from a \code{dwca} object}
\usage{
build(.dwca, file = NULL)
}
\arguments{
\item{.dwca}{A \code{dwca} object}

\item{path}{(Optional) Name of the zip file. Defaults to \code{NULL}, indicating
that the file in question should be saved to a temporary directory.}
}
\value{
Returns an object of class \code{dwca}, with an added \code{path} slot showing
the location of the zip file. However, this function is primarily called for
the side-effect of building a 'Darwin Core Archive' (i.e. a zip file).
}
\description{
Build a Darwin Core Archive from a \code{dwca} object
}
