% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_archive.R
\name{build_archive}
\alias{build_archive}
\title{\code{archive} will attempt to zip a complete Darwin Core archive.
It will call \code{darwin_check} to make sure the data conforms to Darwin Core
standards. It will also need a valid \code{meta.xml} file and \code{eml.xml} file.}
\usage{
build_archive(data, meta = NULL, eml = NULL, folder = ".")
}
\arguments{
\item{data}{A data frame of your data}

\item{meta}{A data frame of your metadata}

\item{eml}{A data frame of your EML metadata}

\item{path}{A path to save the zip file to (root by default)}
}
\value{
No object is returned; this function is called for the side-effect
of building a 'Darwin Core Archive' (i.e. a zip file)
}
\description{
\code{archive} will attempt to zip a complete Darwin Core archive.
It will call \code{darwin_check} to make sure the data conforms to Darwin Core
standards. It will also need a valid \code{meta.xml} file and \code{eml.xml} file.
}
