% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_unique_column.R
\name{build_composite_identifier}
\alias{build_composite_identifier}
\title{Create a composite identifier from two or more columns, separated by a colon.
TODO: Question:: Would a composite identifier column be called occurrenceID?
Should be globally unique - it may be necessary to still add a random number?
Depends on what columns are used to build it.}
\usage{
build_composite_identifier(data, cols = NULL)
}
\arguments{
\item{cols}{character vector of columns to use}
}
\description{
Create a composite identifier from two or more columns, separated by a colon.
TODO: Question:: Would a composite identifier column be called occurrenceID?
Should be globally unique - it may be necessary to still add a random number?
Depends on what columns are used to build it.
}
\keyword{internal}
