% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dwca.R
\name{build_dwca}
\alias{build_dwca}
\title{Build a Darwin Core Archive from a Biodiversity Data Package}
\usage{
build_dwca(pkg = ".", path = NULL)
}
\arguments{
\item{pkg}{A directory containing a Biodiversity Data Package. Defaults to
current working directory.}

\item{path}{(Optional) Path to the resulting zip file. Defaults to \code{NULL},
indicating that the file should be saved to the current parent directory.}
}
\value{
Called exclusively for the side-effect of building a 'Darwin Core
Archive' (i.e. a zip file); doesn't return anything to the workspace.
}
\description{
Similar to \code{devtools::build()}, but to construct a Darwin Core Archive
rather than an R package.
}
\details{
This function looks for specific objects in specific places, namely:
\itemize{
\item Named objects in the \code{data} folder
\item Schema and metadata in the \code{inst} folder
}

If no data are present, or \code{inst/metadata.md} is missing, the function will
terminate with an error. This does not happen if \code{inst/schema.md} is missing,
because this can be built automatically using \code{use_bd_schema()}.
}
