% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_eml.R
\name{build_eml}
\alias{build_eml}
\title{Build and export an EML file using an input markdown template}
\usage{
build_eml(template, output_dir = "./eml.xml", validate = FALSE)
}
\arguments{
\item{template}{string Path to input markdown template}

\item{output_dir}{string Path to output EML file, defaults to current working
directory \code{./eml.xml}}

\item{validate_eml}{logical Validate the EML file using the EML
package function \code{eml_validate()} when TRUE (default: FALSE).}
}
\value{
Writes EML file to output directory
}
\description{
Based on the function \code{set_TextType()} from the EML package.
}
\examples{
\dontrun{
build_eml(template = "./user_template.md")
build_eml(
  template = "./user_template.md",
  output_dir = "./metadata/my_eml.xml",
  validate = TRUE
)
}
}
