% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_identifiers.R
\name{build_identifiers}
\alias{build_identifiers}
\alias{build_random_identifier}
\alias{build_composite_identifier}
\title{Create a random identifier column}
\usage{
build_random_identifier(data, random_seed = NULL)

build_composite_identifier(data, cols = NULL)
}
\arguments{
\item{data}{A tibble}

\item{random_seed}{Integer: seed value (optional)}

\item{cols}{character vector of columns to use}
}
\description{
Create a random identifier column

Create a composite identifier from two or more columns, separated by a colon.
TODO: Question:: Would a composite identifier column be called occurrenceID?
Should be globally unique - it may be necessary to still add a random number?
Depends on what columns are used to build it.
}
