% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_unique_column.R
\name{build_random_identifier}
\alias{build_random_identifier}
\title{Create a random identifier column}
\usage{
build_random_identifier(data, random_seed = NULL)
}
\arguments{
\item{data}{A tibble}

\item{random_seed}{Integer: seed value (optional)}
}
\description{
Create a random identifier column
}
\keyword{internal}
