% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_schema.R
\name{build_schema}
\alias{build_schema}
\title{Create a \code{schema} for a Darwin Core Archive}
\usage{
build_schema(directory = "data")
}
\arguments{
\item{directory}{A folder containing Darwin Core data. Defaults to \code{data}.}
}
\value{
Does not return an object to the workspace; called for the side
effect of building a file named \code{meta.xml} in the specified directory.
}
\description{
A schema is an xml document that maps the files and field names in a DwCA.
It works by detecting column names on csv files in a specified directory;
these should all be Darwin Core terms for this function to produce reliable
results.
}
