% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\alias{check_local}
\alias{check_remote}
\title{Check a \code{dwca} for inconsistencies with the standard}
\usage{
check(.dwca, type = c("local", "remote"))

check_local(.dwca, max_n = NULL)

check_remote(.dwca = NULL, file = NULL)
}
\arguments{
\item{.dwca}{An object of class \code{dwca}, created using \code{dwca()}}

\item{max_n}{Optional limit to the number of failed tests allowed}
}
\description{
Check a \code{dwca} for inconsistencies with the standard
}
\details{
\code{type = "remote"} uses the ALA \code{validate} API, which in turn calls
\code{dwca-validator} Python. \code{type = "local"} duplicates this functionality
via \code{{pointblank}}
}
