% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_archive.R, R/get_report.R,
%   R/print_gbif_validator.R
\name{check_archive}
\alias{check_archive}
\alias{get_report}
\alias{print.gbif_validator}
\title{Check whether an archive meets the Darwin Core Standard via API}
\usage{
check_archive(wait = TRUE, quiet = FALSE)

get_report(key, wait = TRUE, quiet = FALSE)

\method{print}{gbif_validator}(x, n = 5, ...)
}
\arguments{
\item{key}{A key that uniquely identifies your query}

\item{x}{An object to print.}

\item{n}{Number of entries to print per file}

\item{\dots}{Additional arguments, currently ignored.}
}
\value{
Both \code{\link[=check_archive]{check_archive()}} and \code{\link[=get_report]{get_report()}} return an object of class
\code{gbif_validator} to the workspace, invisibly if \code{quiet = TRUE}.
}
\description{
Check whether a specified Darwin Core Archive is ready for
sharing and publication, according to the Darwin Core Standard.
\code{\link[=check_archive]{check_archive()}} tests the specified archive using an online validation
service by sending the archive via an API and returning the results.
Currently only supports validation using GBIF.
}
\details{
\code{\link[=check_archive]{check_archive()}} both \code{POST}s the specified archive to the GBIF validator
API and then calls \code{\link[=get_report]{get_report()}} to retrieve (\code{GET}) the result.
\code{\link[=get_report]{get_report()}} is exported to allow the user to download results at a later
time should they wish; this is more efficient than repeatedly generating
queries with \code{\link[=check_archive]{check_archive()}} if the underlying data are unchanged. A third
option is simply to assign the outcome of \code{\link[=check_archive]{check_archive()}} or \code{\link[=get_report]{get_report()}}
to an object then call it again, which uses \code{\link[=print.gbif_validator]{print.gbif_validator()}} to
format the result nicely. This approach doesn't require any further API calls
and as such is considerably faster.

Note that information returned by these functions is provided verbatim from
the institution API, not from galaxias.
}
\examples{
\dontrun{
# add GBIF login details
galaxias_config(gbif = list(username = "your-gbif-username",
                            email = "your-gbif-email",
                            password = "your-gbif-password"))

# Check archive against Darwin Core Standard criteria
check_archive("dwc-archive.zip")
}
}
\seealso{
\code{\link[=check_directory]{check_directory()}} which runs checks on a folder directory locally,
rather than via API.
}
