% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dwc_df.R
\name{check_dwc}
\alias{check_dwc}
\alias{check_fields}
\alias{check_basisOfRecord}
\title{Check for Darwin Core field conformance}
\usage{
check_dwc(df)

check_fields(df)

check_basisOfRecord(df, level = c("inform", "warn", "abort"))
}
\arguments{
\item{df}{a tibble containing data}

\item{level}{what action should the function take for non-conformance?
Defaults to \code{"inform"}.}
}
\description{
Function to check whether a data.frame or tibble conforms to DwC standards
NOTE: Option to use \code{bdc} for checks, and/or \code{pointblank} for running them
}
