% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dwca.R
\name{check_dwca}
\alias{check_dwca}
\title{pointblank-based checking of a \code{dwca} object}
\usage{
check_dwca(.dwca, max_n = NULL)
}
\arguments{
\item{.dwca}{An object of class \code{dwca}, created using \code{dwca()}}

\item{max_n}{Optional limit to the number of failed tests allowed}
}
\description{
Description to follow.
}
