% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eml.R
\name{check_eml}
\alias{check_eml}
\title{Check for required ALA fields in an EML list}
\usage{
check_eml(eml_list)
}
\arguments{
\item{eml_list}{list? of EML fields}
}
\value{
Console output
}
\description{
Check for required ALA fields in an EML list
}
