% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_fields.R
\name{check_fields}
\alias{check_fields}
\title{\code{Check_fields()} wrapper function that calls all other \verb{check_*()} functions}
\usage{
check_fields(data, all_fields = TRUE)
}
\arguments{
\item{data}{A tibble}

\item{all_fields}{A logical indicating whether to check all fields, or only
required fields. Default is \code{TRUE}.}
}
\description{
\code{Check_fields()} wrapper function that calls all other \verb{check_*()} functions
}
