% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_metadata}
\alias{check_metadata}
\title{Check validity of a metadata statement}
\usage{
check_metadata(filename)
}
\arguments{
\item{filename}{A local file containing metadata in xml format. This will
typically be called \code{eml.xml} and can be constructed from markdown using
\code{build_metadata()}.}
}
\value{
Invisibly returns a tibble showing parsed errors; or an empty
tibble if no errors are identified.
}
\description{
In the Darwin Core standard, metadata statements are mandatory and must be
provided in Ecological Meta Language (EML) in a file called \code{eml.xml}. This
function applies a series of checks designed by GBIF to check the structure
of the specified \code{filename} for consistency with the standard. Note,
however, that this function doesn't check the \emph{content} of those files,
meaning a file could be structurally sound and still be lacking critical
information.
}
\details{
This function uses local versions of \code{dc.xsd}, \code{eml-gbif-profile.xsd} and
\code{eml.xsd} downloaded from
\link{http://rs.gbif.org/schema/eml-gbif-profile/1.3/} on 2024-09-25.
}
