% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_occurrences.R
\name{check_occurrences}
\alias{check_occurrences}
\title{Check occurrence data for Darwin Core conformance}
\usage{
check_occurrences(.df)
}
\arguments{
\item{.df}{A tibble against which checks should be run}
}
\value{
Invisibly returns the input, but primarily called for the
side-effect of running check functions on that input.
}
\description{
Function to check whether a \code{data.frame} or \code{tibble} conforms to Darwin
Core standards. While most users will only want to call \code{suggest_workflow()},
the underlying check functions are exported for detailed work, or for
debugging.
}
