% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_fields.R
\name{check_recommended_fields}
\alias{check_recommended_fields}
\title{Check for presence of only the recommended fields in DwC}
\usage{
check_recommended_fields(data)
}
\description{
Check for presence of only the recommended fields in DwC
}
