% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_fields.R
\name{check_unique_identifiers}
\alias{check_unique_identifiers}
\title{Check for presence of a unique identifier field required by DwC}
\usage{
check_unique_identifiers(data)
}
\arguments{
\item{data}{A tibble}
}
\description{
The \code{check_unique_identifiers()} function is used to check for the presence
of any of the three unique identifier fields required by DwC (\code{occurrenceID},
\code{catalogueNumber}, \code{recordNumber}). If none are found, the function will call
\code{detect_unique_column()}, which will attempt to detect a unique column in the
data that may be renamed, or else assist the user in creating one.
}
