% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{column_matching}
\alias{column_matching}
\title{User supplied column mapping to DwC terms in two column format, first column
user term, second column is the matching DwC term}
\usage{
column_matching(data, mapping)
}
\description{
User supplied column mapping to DwC terms in two column format, first column
user term, second column is the matching DwC term
}
