% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.R
\name{convert_md_tibble_to_list}
\alias{convert_md_tibble_to_list}
\title{Convert metadata tibble into list}
\usage{
convert_md_tibble_to_list(metadata_tibble)
}
\arguments{
\item{metadata_tibble}{}
}
\value{
named list
}
\description{
Convert metadata tibble into list
}
