% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bd_project.R
\name{create_bd_project}
\alias{create_bd_project}
\title{Create a new Biodiversity Data Project}
\usage{
create_bd_project(
  path,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive()
)
}
\arguments{
\item{path}{A path to the directory to create}

\item{rstudio}{(logical) should we create an RStudio project?}

\item{open}{(logical) should the resulting project be opened?}
}
\description{
Sets up a 'Biodiversity Data Project', which is a repository for
storing biodiversity data, ideally according to the Darwin Core Standard.
In practice, this is a modified version of \code{usethis::create_project()},
and calls that function internally.
}
