% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_package.R
\name{create_data_package}
\alias{create_data_package}
\title{Function to create a new RStudio Project}
\usage{
create_data_package(path, copyright_holder = NULL)
}
\arguments{
\item{path}{A path to the directory to create}
}
\description{
Sets up all 'Biodiversity Data Package', which is a modified R package for
building Darwin Core Archives. Based heavily on usethis::create_tidy_package
}
