% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{darwin_check}
\alias{darwin_check}
\title{Primary function to check (and optionally conform) a data frame with Darwin
Core standards, and return a summary report.}
\usage{
darwin_check(data, mend, all_fields = TRUE)
}
\arguments{
\item{data}{A data frame of biological observations}

\item{mend}{A logical indicating whether to attempt to fix the data, FALSE by
default. This should be interactive (\code{build_} and \code{mend_} functions should be
called).}

\item{all_fields}{A logical indicating whether to check all fields, or only
required DwC fields. Default is \code{TRUE}.}
}
\value{
A report of the data columns that match Darwin Core terms. If \code{mend}
is TRUE, a modified data frame will be returned depending on user input.
}
\description{
Generate a report of the data columns that match Darwin Core terms,
summarising the compliance of the data with Darwin Core standards, and
recommended fixes where required. Some common issues can be corrected
interactively if \code{mend} is TRUE (experimental). Otherwise, a user can
manually fix their data, and run this function again to check if the data
conforms. against Darwin Core required and recommended (optional) fields.
Some common issues can be corrected when prompted.
}
