% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_column_names.R
\name{detect_column_names}
\alias{detect_column_names}
\title{Detects DarwinCore column names}
\usage{
detect_column_names(data)
}
\arguments{
\item{data}{a \code{tibble} containing biological observations}
}
\value{
A tibble identical to the input, except with 'corrected' column
names.
}
\description{
This function takes a tibble of user-supplied data, and seeks to match
supplied column names to those given in the DarwinCore schema.
}
\details{
This function is in progress and behaviour may change. At present it...
}
