% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_unique_column.R
\name{detect_unique_column}
\alias{detect_unique_column}
\title{Detect if a unique-like column is found in the data}
\usage{
detect_unique_column(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
Based on user choice, 1) input tibble with a renamed column, 2) input
tibble with a new column, or 3) NULL
}
\description{
Unique columns are detected (every row is unique). If yes, the user is given
the option to rename it. If no unique column is detected, the user is given
the option to create one. A unique identifier column is required to comply
with DwC standard.
}
