% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_metadata.R
\name{edit_metadata}
\alias{edit_metadata}
\title{Open a markdown template file for creating EML metadata}
\usage{
edit_metadata(path = ".", type = "ALA", licence = 1)
}
\arguments{
\item{path}{string Directory where metadata template will be copied to,
defaults to current working directory.}

\item{type}{string Which template flavour to use:
\itemize{
\item ALA (default)
\item More to come...
}}

\item{licence}{numeric Which licence to use for the dataset:
\itemize{
\item 1 = Creative Commons Zero - CC0
\item 2 = Attribution – CC BY (4.0)
\item 3 = Attribution-Noncommercial – CC BY-NC
}
See
\link{article}(https://support.ala.org.au/support/solutions/articles/
6000197133-what-licence-should-i-use-) for more information.}
}
\value{
The template file is opened in RStudio.
}
\description{
Open a markdown template file for creating EML metadata
}
