% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_str.R
\name{eml_template_ls}
\alias{eml_template_ls}
\title{ALA eml.xml template}
\usage{
eml_template_ls()
}
\value{
Nested list structure
}
\description{
ALA eml.xml template
}
\examples{
eml <- eml_template_ls()
}
\keyword{internal}
