% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.R
\name{extract_md_contents}
\alias{extract_md_contents}
\title{Extract metadata information and wrangle into tibble}
\usage{
extract_md_contents(scan_output)
}
\arguments{
\item{scan_output}{output of \code{read_metadata_md()}}
}
\value{
tibble where text from each field is concatenated by header
}
\description{
Extract metadata information and wrangle into tibble
}
\examples{
scan_output <- read_metadata_md("doc/westerband_template.md")
extract_md_contents(scan_output)
}
