% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_str.R
\name{extract_name_from_ls}
\alias{extract_name_from_ls}
\title{Extract first or last name from Name field}
\usage{
extract_name_from_ls(list, var = "Name", which)
}
\arguments{
\item{list}{output from \code{convert_md_tibble_to_list()}}

\item{var}{character, name field (Name/Creator)}

\item{which}{character denoting which part of name to extract (first/last)}
}
\value{
vector of length 1 containing first or last name
}
\description{
Extract first or last name from Name field
}
\examples{
convert_md_tibble_to_list(metadata_tibble) |> extract_name_from_ls(which = "last")
}
\keyword{internal}
