% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{galaxias-dplyr}
\alias{galaxias-dplyr}
\alias{mutate.dwc_df}
\alias{rename.dwc_df}
\alias{rename_with.dwc_df}
\alias{select.dwc_df}
\title{Functions masked from \code{dplyr}}
\usage{
\method{mutate}{dwc_df}(
  .data,
  ...,
  .by = NULL,
  .keep = c("all", "used", "unused", "none"),
  .before = NULL,
  .after = NULL
)

\method{rename}{dwc_df}(.data, ...)

\method{rename_with}{dwc_df}(.data, .fn, .cols = everything(), ...)

\method{select}{dwc_df}(.data, ...)
}
\description{
These functions are masked for class \code{dwc} to support type-specific data
checks. Experimental.
}
