% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxias-package.R
\docType{package}
\name{galaxias-package}
\alias{galaxias}
\alias{galaxias-package}
\title{Building Darwin Core Archives for the Atlas of Living Australia}
\description{
The Darwin Core Archive (DwCA) is a transferable unit of
biodiversity data and associated metadata that allows facilities
like the Atlas of Living Australia (ALA) to efficiently integrate and share
openly with the community. Creating DwCAs by hand is a notoriously fiddly,
task. 'galaxias' enables users to efficiently reformat their data to Darwin
Core Standards and generate production-ready DwCA in a reproducible and
transparent workflow.

\code{galaxias} is named for a genus of Australian freshwater fish.
}
\section{Object classes}{

\code{galaxias} supports the following object classes:
\itemize{
\item \code{dwc_a} Darwin Core Archive objects
\item \code{dwc_df} data frames to be checked against the Darwin Core standard
\item \code{dwc_xml} xml objects to be checked against the Darwin Core standard
}

All can be created using the \code{galaxias()} function
}

\section{Functions}{

\strong{Setting up an archive}
\itemize{
\item \code{\link[=dwca]{dwca()}} create pipable \code{dwca} objects
\item \code{\link[=add_occurrences]{add_occurrences()}}, \code{\link[=add_events]{add_events()}}, \code{\link[=add_multimedia]{add_multimedia()}}, \code{\link[=add_metadata]{add_metadata()}} to add data to \code{dwca} objects
}

\strong{Data handling}
\itemize{
\item \code{\link[=build_random_identifier]{build_random_identifier()}}, \code{\link[=build_composite_identifier]{build_composite_identifier()}} uuids
\item \code{\link[=get_metadata_template]{get_metadata_template()}}, \code{\link[=get_metadata_example]{get_metadata_example()}} get example metadata in xml
\item \code{\link[=read_md]{read_md()}}, \code{\link[=write_md]{write_md()}} to convert between markdown and xml
}

\strong{Checking and publication}
\itemize{
\item \code{\link[=build]{build()}} build a \code{dwca} into a zip file
\item \code{\link[=check]{check()}} check a \code{dwca} object locally or via API
\item \code{\link[=publish]{publish()}} send your data to the ALA for publication
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Fonti Kar \email{fonti.kar@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
  \item Andrew Schwenke \email{andrew.schwenke@csiro.au}
}

}
\keyword{internal}
