% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxias.R
\name{galaxias}
\alias{galaxias}
\title{Use \code{galaxias}-specific object classes}
\usage{
galaxias(...)
}
\arguments{
\item{type}{(string) currently only "occurrence-core" is accepted}
}
\value{
If no arguments are supplied, returns an object of class \code{dwc_a}
(Darwin Core Archive). Otherwise returns \code{dwc_df} for a data.frame or tibble,
and \code{dwc_xml} for xml documents.
}
\description{
The \code{galaxias} package uses an object-oriented framework. The \code{galaxias()}
function is a lightweight way to access those frameworks for efficient usage.
The most straightforward way to achieve this is to start a pipe for building,
checking and publishing a Darwin Core Archive; but you can also place
\code{galaxias()} \emph{within} a pipe to unlock DwC-checking behaviour in associated
tibbles or xml documents.
}
