% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxias_project.R
\name{galaxias_project}
\alias{galaxias_project}
\alias{galaxias_package}
\title{Create a new \code{galaxias} project}
\usage{
galaxias_project(path, ...)

galaxias_package(path, ...)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{\dots}{Other arguments passed to \code{usethis::create_project()} or
\code{usethis::create_package()}}
}
\value{
Does not return an object to the workspace; called for the
side-effect of building a new project or package.
}
\description{
This is a wrapper to \code{usethis} functions \code{create_project()} and
\code{create_package()}. It differs in that it builds the requisite folders
\code{"data-raw"} for 'initial' data, \code{"data"} for processed data, and
\code{"data-published"} for data that has been
formatted according to the Darwin Core Standard.
}
