% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report}
\alias{generate_report}
\title{Report generation}
\usage{
generate_report(data)
}
\arguments{
\item{data}{data frame to be used in the report}
}
\description{
Understanding the logging and report behaviours and edge cases:
The idea of the log environment is that a user can run individual functions
within a session, and then decide to print a report at the end of their
session. As the list is in an env, we can access it anywhere. The downside is
it can be a little more difficult to manage and debug if issues crop up. The
alternative approach is we only allow a report to be generated at
the end of complete pipeline run. The advantage of this E2E approach would be
easier handling of logs - just passing a named list through the pipeline
appending logs, which would probably be more reliable and easier to manage.
}
\details{
Notes on log env behaviours (in progress):
\itemize{
\item entries are overwritten
\itemize{
\item run validate latitude twice, the first log entry will be overwritten
}
\item Using \code{load_all()} clears the log_env because of \code{.onLoad()}
\item Custom output directory possible but requires extra steps
}
}
