% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_blank_metadata}
\alias{get_blank_metadata}
\alias{get_example_metadata}
\title{Get example metadata from ALA}
\usage{
get_blank_metadata()

get_example_metadata(search = NULL, id = NULL)
}
\arguments{
\item{search}{A string to search using \code{galah::search_all_collections()}}

\item{id}{An ALA data resource number}
}
\description{
Wrapper function to show some example metadata
}
