% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get example metadata from ALA}
\usage{
get_metadata(search = NULL, id = NULL, file = "metadata.rmd")
}
\arguments{
\item{search}{A string to search using \code{galah::search_all_collections()}}

\item{id}{An ALA data resource number}

\item{file}{File name to save the resulting markdown}
}
\description{
Wrapper function to show some example metadata
}
