% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata_template}
\alias{get_metadata_template}
\alias{get_metadata_example}
\title{Get example metadata from ALA}
\usage{
get_metadata_template()

get_metadata_example(search = NULL, id = NULL)
}
\arguments{
\item{search}{A string to search using \code{galah::search_all_collections()}}

\item{id}{An ALA data resource number}
}
\description{
Wrapper function to show some example metadata
}
