% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofish.R
\name{gofish}
\alias{gofish}
\alias{darwin_core_archive}
\alias{print.darwin_core_archive}
\title{Start a pipe to build a Darwin Core Archive}
\usage{
gofish()

darwin_core_archive()

\method{print}{darwin_core_archive}(x, ...)
}
\arguments{
\item{type}{(string) currently only "occurrence-core" is accepted}

\item{extentions}{Not currently supported; what further extentions should be added.
These are additional files such as 'events' or 'media' that are additional
to occurrences}
}
\value{
An object of class \code{dwca} containing three slots, corresponding to
the required objects for a Darwin Core Archive.
}
\description{
\code{galaxias} uses an object-oriented framework. This starts the pipe to build
the object in question. \code{gofish()} and \code{darwin_core_archive()} are identical
in function, but not in string length.
}
