% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_table.R
\name{import_map}
\alias{import_map}
\title{Import an existing lookup table for DwC terms}
\usage{
import_map(mapping, index, dataset = NULL)
}
\arguments{
\item{mapping}{data frame Containing at least two columns}

\item{index}{numeric vector of length 2. The first element is the column
index to use for the data column names, the second element is the column
index to use for the DwC terms e.g. \code{index = c(1, 2)}.}

\item{dataset}{data frame Optional dataset to which the lookup table
pertains, used for validation.}
}
\description{
Imports a mapping table, ensuring dataset column names align with DwC terms.
\itemize{
\item Requires a data frame with at least two columns:
\itemize{
\item First: Column names of your dataset.
\item Second: Corresponding DwC terms.
}
\item Each row must correctly pair a dataset column name with a DwC term, e.g.,
\code{data.frame("sci name", "scientificName")}.
}
}
\section{dev notes}{

In progress, potential use cases:
\itemize{
\item meta xml mapping process
\item E2E use; pipeline w/ existing lookup table
\item Importing a lookup table created in a previous session (with \code{map2dwc()})
}
}

