% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meta_xml.R
\name{make_core_xml}
\alias{make_core_xml}
\title{Create core subfield}
\usage{
make_core_xml(x, file_name = "occurrence.csv")
}
\arguments{
\item{x}{output of Darwinised occurrence data, output from \code{detect_dwc_columns()}}

\item{file_name}{file name of DwC standard occurrence data}
}
\value{
xml document
}
\description{
Create core subfield
}
\examples{
DwC_occurrence_data <- tibble(field_names =  c("basisOfRecord", "eventDate", "decimalLatitude"))
make_core_xml(DwC_occurrence_data)
}
