% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_table.R
\name{map_fields}
\alias{map_fields}
\title{Map User Columns to Darwin Core Terms}
\usage{
map_fields(data)
}
\arguments{
\item{data}{data frame Containing a dataset of which to map DwC
terms to.}
}
\value{
data frame With renamed columns based on the mapping. Exiting the
mapping early (by using the \code{exit} command when prompted) will return a data
frame with any mappings that were made prior to exiting.
}
\description{
For now, mappings are used to rename columns. Can be modified or extended to
output a dictionary for the meta.xml and preserve column names etc.

This interactive function assists in mapping user data columns to the
standard Darwin Core (DwC) terms. It suggests potential DwC matches for each
column name in the user's dataset. Suggestions are generated based on word
component matches and string distance metrics to provide the best possible
matches.

Suggestions are presented to the user in a numbered list, and the user is
then prompted to select the correct DwC term. If no correct match was found,
the user is prompted to manually input the correct term if known, or skip the
column, before continuing to the next column.

Note that a minimum of five matches is always returned, even if they are
only low confidence matches.
}
