% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tibble_to_list.R
\name{md_tibble_to_list}
\alias{md_tibble_to_list}
\title{Convert metadata tibble into list}
\usage{
md_tibble_to_list(metadata_tibble)
}
\arguments{
\item{metadata_tibble}{output from \code{md_to_tibble()}}
}
\value{
named list with field name as name
}
\description{
Convert metadata tibble into list
}
\examples{
path_to_md_template <- system.file("markdown/westerband_template.md", package = "galaxias")
scan_output <- read_metadata_md(path_to_md_template)
metadata_tibble <- extract_md_contents(scan_output)
metadata_ls <- metadata_tibble |> md_tibble_to_list()
}
