% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_to_tibble.R
\name{md_to_tibble}
\alias{md_to_tibble}
\title{Extract metadata text and wrangle into tibble}
\usage{
md_to_tibble(scan_output)
}
\arguments{
\item{scan_output}{output of \code{read_metadata_md()}}
}
\value{
tibble where text from each field is concatenated by header
}
\description{
Extract metadata text and wrangle into tibble
}
\examples{
scan_output <- read_metadata_md("doc/westerband_template.md")
extract_md_contents(scan_output)
}
