% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\alias{parse_md_to_tibble}
\alias{parse_md_to_list}
\alias{parse_md_to_xml}
\alias{parse_tibble_to_list}
\alias{parse_tibble_to_xml}
\alias{parse_list_to_xml}
\alias{parse_xml_to_list}
\title{Parse between common object types}
\usage{
parse_md_to_tibble(x)

parse_md_to_list(x)

parse_md_to_xml(x)

parse_tibble_to_list(df)

parse_tibble_to_xml(df)

parse_list_to_xml(x)

parse_xml_to_list(x)
}
\arguments{
\item{x}{For markdown parsers, a vector of class \code{character}, such as
imported using \code{base::readLines()}, containing text in markdown format. For
other functions, an object of class \code{list} or \code{xml_document}.}

\item{df}{A \code{tibble} built within \code{md_to_tibble()}}
}
\description{
It is important in \code{galaxias} to be able to move between \code{tibble}s, \code{list}s,
and \code{xml} documents. Primarily called by \code{read_md()} and \code{write_md()}, these
functions are exported for clarity and debugging purposes.
}
\details{
In practice, the sequence of transformations is fixed, being \code{string} >
\code{tibble} > \code{list} > \code{xml}. Nonetheless, it can be useful sometimes to have
helper functions that handle this hierarchy 'under the hood'.

Note that \code{parse_tibble_to_list()} is a recursive function, and is highly
bespoke to tibbles of the type returned by \code{parse_md_to_tibble()}.
\code{parse_list_to_xml()} is synonymous with \code{xml2::as_xml_document()} and
\code{parse_xml_to_list()} is synonymous with \code{xml2::as_list()}
}
