% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dwc_a}
\alias{print.dwc_a}
\alias{print.dwc_df}
\alias{print.dwc_xml}
\title{print functions for \code{galaxias} objects}
\usage{
\method{print}{dwc_a}(x, ...)

\method{print}{dwc_df}(x, width = NULL, ..., n = NULL)

\method{print}{dwc_xml}(x, ...)
}
\description{
There are three of these
}
