% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_API}
\alias{print_API}
\alias{print.gbif_validator_response}
\title{Print objects returned by APIs}
\usage{
\method{print}{gbif_validator_response}(x, n = 5, ...)
}
\arguments{
\item{x}{An object to print.}

\item{n}{Number of entries to print}

\item{\dots}{Additional arguments, currently ignored.}
}
\description{
Currently only \code{validate_archive()}
}
