% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_gbif_validator.R
\name{print_validation}
\alias{print_validation}
\alias{print.gbif_validator_post}
\alias{print.gbif_validator_response}
\title{Print objects returned by APIs}
\usage{
\method{print}{gbif_validator_post}(x, ...)

\method{print}{gbif_validator_response}(x, n = 5, ...)
}
\arguments{
\item{x}{An object to print.}

\item{\dots}{Additional arguments, currently ignored.}

\item{n}{Number of entries to print}
}
\description{
Currently only \code{validate_archive()} and \code{get_validator_report()}, where the
latter is called by the former.
}
