% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dwca.R, R/write_dwca.R
\name{read_dwca}
\alias{read_dwca}
\alias{write_dwca}
\title{Function to read an existing Darwin Core Archive}
\usage{
read_dwca(file)

write_dwca(.dwca, file = NULL)
}
\arguments{
\item{.dwca}{A \code{dwca} object}

\item{path}{(Optional) Name of the zip file. Defaults to \code{NULL}, indicating
that the file in question should be saved to a temporary directory.}
}
\value{
Returns an object of class \code{dwca}, with an added \code{path} slot showing
the location of the zip file. However, this function is primarily called for
the side-effect of building a 'Darwin Core Archive' (i.e. a zip file).
}
\description{
Early stages
}
