% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mutate}
\alias{rename}
\alias{rename_with}
\alias{select}
\alias{as_tibble}
\alias{everything}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:rename]{rename_with}}, \code{\link[dplyr]{select}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidyselect}{\code{\link[tidyselect]{everything}}}
}}

