% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_camel_case}
\alias{rename_camel_case}
\alias{rename_identifier}
\title{Rename columns to DwC-compatible types}
\usage{
rename_camel_case(data)

rename_identifier(.data, .col)
}
\arguments{
\item{data}{A tibble}

\item{.col}{What column should be renamed?}
}
\value{
A tibble
}
\description{
These functions are in prep
}
