% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_unique_column.R
\name{rename_interactive}
\alias{rename_interactive}
\title{Rename a column interactively.}
\usage{
rename_interactive(data, unique_columns)
}
\arguments{
\item{data}{A tibble}

\item{unique_columns}{A character vector of column names}
}
\value{
A tibble with renamed columns
}
\description{
This is called by \code{detect_unique_column} when one or more unique columns are
detected. The user is then given the option to rename the column to one of
the Darwin Core standard identifiers, if it is appropriate to do so.
}
\keyword{internal}
