% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_table.R
\name{split_into_words}
\alias{split_into_words}
\title{Split a string into words}
\usage{
split_into_words(string)
}
\arguments{
\item{string}{string}
}
\value{
character vector
}
\description{
Splits on non-alphanumeric characters/spaces and handles camel case by
inserting spaces before uppercase letters following lowercase ones. Properly
handles acronyms: splits 'languageEnglish', 'languageEN', but not 'EN'.
}
\seealso{
Called by \code{\link[=word_overlap_score]{word_overlap_score()}}
}
\keyword{internal}
