% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_workflow.R
\name{suggest_workflow}
\alias{suggest_workflow}
\title{Suggest workflow to make data compliant with Darwin Core}
\usage{
suggest_workflow(.df)
}
\arguments{
\item{.df}{A tibble against which checks should be run}
}
\value{
Invisibly returns the input, but primarily called for the
side-effect of running check functions on that input.
}
\description{
Function to check whether a \code{data.frame} or \code{tibble} conforms to Darwin
Core standards. While most users will only want to call \code{suggest_workflow()},
the underlying check functions are exported for detailed work, or for
debugging.
}
