% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{testthat}
\alias{testthat}
\alias{expect_exists}
\title{testthat functions}
\usage{
expect_exists(path)
}
\arguments{
\item{path}{(string) path to a file to test for existence}
}
\description{
These extend \code{testthat} for use in \code{galaxias}. Experimental
}
