% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_to_dwc.R
\name{tibble_to_dwc}
\alias{tibble_to_dwc}
\title{Convert a tibble to Darwin-Core compatible content and colnames}
\usage{
tibble_to_dwc(data)
}
\arguments{
\item{data}{a \code{tibble} containing biological observations}
}
\value{
A tibble identical to the input, except with 'corrected' column
names.
}
\description{
This function takes a tibble of user-supplied data, and seeks to match
supplied column names to those given in the DarwinCore schema. It also
corrects common problems and removes non-DwC-compatible fields.
For finer control, see \code{check()}.
}
\details{
This function is in progress and behaviour may change. At present it...
}
