% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_str.R, R/update_eml_field.R
\name{update_title}
\alias{update_title}
\title{Update Title field}
\usage{
update_title(template, metadata_ls)

update_title(template, metadata_ls)
}
\arguments{
\item{template}{ALA template structure generated by \code{eml_template_ls}}

\item{metadata_ls}{Metadata list generated by \code{convert_md_tibble_to_list}}
}
\description{
Update Title field

Update Title field in list template
}
\examples{
eml <- eml_template_ls()
eml
update_title(eml,metadata_ls)
eml <- eml_template_ls()
eml
update_title(eml,metadata_ls)
}
