% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_basisOfRecord.R
\name{use_basisOfRecord}
\alias{use_basisOfRecord}
\title{Add a \code{basisOfRecord} field to a \code{tibble}}
\usage{
use_basisOfRecord(.df, value)
}
\arguments{
\item{.df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}

\item{value}{what value should this field take? Only accepts \code{camelCase}, for
consistency with field names.}
}
\value{
A tibble with an attached \code{basisOfRecord} field
}
\description{
basisOfRecord is a standard - often required - field in the Darwin Core
Standard, which describes broadly how the data were collected.
}
\details{
Currently only accepts a single argument. Could probably be made more
flexible.
}
