% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_citation}
\alias{use_bd_citation}
\title{Add \code{CITATION} file to a biodiversity data repository}
\usage{
use_bd_citation()
}
\description{
Add \code{CITATION} file to a biodiversity data repository
}
