% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_data}
\alias{use_bd_data}
\title{Add \code{data} folder to a biodiversity data project}
\usage{
use_bd_data(..., overwrite = FALSE)
}
\arguments{
\item{...}{Unquoted names of an object to save.}

\item{overwrite}{(logical) Should existing objects be overwritten? Defaults
to FALSE.}
}
\description{
This function places specified objects in the \code{data} folder as \code{.csv} files.
Note that this is very different from \code{usethis::use_data()} which uses \code{.rda}
format (and then only when \code{internal = FALSE}). This is provided for
consistency with \code{usethis}, but a more flexible approach is simply to use
\code{readr::write_csv()}.
}
