% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_data_raw}
\alias{use_bd_data_raw}
\title{Add data folders to a Biodiversity Data Package}
\usage{
use_bd_data_raw()
}
\description{
Add a script to \code{data-raw} with example code of how to rename/select/relocate
fields, then save data out to \code{data} folder
}
