% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_readme_rmd}
\alias{use_bd_readme_rmd}
\title{Add \code{README.Rmd} to a biodiversity data repository}
\usage{
use_bd_readme_rmd()
}
\description{
This function adds \code{galaxias}-specific \code{README.Rmd} instead of the \code{usethis}
default.
}
\details{
FIXME: add function use_bd_readme_md
}
