% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_schema}
\alias{use_bd_schema}
\title{Add \code{schema.md} to a Biodiversity Data Package}
\usage{
use_bd_schema()
}
\description{
Builds a file called \code{schema.md} in the \code{inst} folder; this folder is
created if not already present. The schema is a standard part of Darwin Core
Archives and requires data to be present in the \code{data} folder to do anything
useful.
}
