% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_testthat}
\alias{use_bd_testthat}
\title{Add Darwin Core tests to a biodiversity data repository}
\usage{
use_bd_testthat()
}
\description{
This optional first initiates the \code{tests/testthat} folder, then adds a set of
boilerplate tests suitable for checking data. These tests all assume that
data has been added to the \code{data} folder following the steps outlined in
\code{data_manipulation_script.R} (created using \code{use_bd_data_raw()}).
}
