% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_vignette}
\alias{use_bd_vignette}
\title{Add a vignette to report on contents of a biodiversity data repository}
\usage{
use_bd_vignette()
}
\description{
Still in test, this function builds a report-style vignette so you (and your
users) can see what kind of data is in your package.
}
