% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_collection.R
\name{use_collection}
\alias{use_collection}
\title{Add museum- or collection-specific information to a \code{tibble}}
\usage{
use_collection(
  .df,
  datasetID = NULL,
  datasetName = NULL,
  catalogNumber = NULL,
  .keep = "unused"
)
}
\arguments{
\item{datasetID}{An identifier for the set of data. May be a global unique
identifier or an identifier specific to a collection or institution.}

\item{datasetName}{The name identifying the data set from which the record
was derived.}

\item{catalogNumber}{A unique identifier for the record within the data set
or collection.}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{dplyr::mutate}, which defaults to \code{"all"} this defaults to
\code{"unused"}; i.e. only keeps Darwin Core fields, and not those fields used to
generate them.}

\item{df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}
}
\value{
A tibble with the requested fields added.
}
\description{
Format fields that specify the collection or catelog number of a
specimen or occurrence record.

In practice this is no different from using \code{mutate()}, but gives some
informative errors, and serves as a useful lookup for fields in
the Darwin Core Standard.
}
\details{
Examples of \code{datasetID} values:
\itemize{
\item \verb{b15d4952-7d20-46f1-8a3e-556a512b04c5}
}

Examples of \code{datasetName} values:
\itemize{
\item \verb{Grinnell Resurvey Mammals}
\item \verb{Lacey Ctenomys Recaptures}
}

Examples of \code{catalogNumber} values:
\itemize{
\item \code{145732}
\item \verb{145732a}
\item \code{2008.1334}
\item \code{R-4313}
}
}
