% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_countryCode.R
\name{use_countryCode}
\alias{use_countryCode}
\title{Add a \code{countryCode} field to a \code{tibble}}
\usage{
use_countryCode(df, value = NULL)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}

\item{value}{what value should this field take? Only accepts camelCase, for
consistency with field names.}
}
\value{
A tibble with an attached \code{countryCode} field
}
\description{
\code{countryCode} is a Darwin Core field that defines the country of observation.
\href{https://dwc.tdwg.org/list/#dwc_countryCode}{TDWG recommend} that
\code{countryCode} should be supplied according to the
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 ALPHA-2}
standard, and that is enforced here.
}
\details{
Currently only accepts a single argument. Could probably be made more
flexible. Also checks whether \code{country} field is present, and if so,
whether its' contents match those given in \code{value}; if not returns a
\code{warning}.
}
