% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_data_description}
\alias{use_data_description}
\title{Add DESCRIPTION to a Biodiversity Data Package}
\usage{
use_data_description()
}
\description{
add \code{usethis} to SUGGESTS (not DEPENDS, because code is only in data-raw)
add any packages mentioned in \code{data_manipulation_script.R} to SUGGESTS
}
