% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_data.R
\name{use_data_events}
\alias{use_data_events}
\title{Save standardised Event-based data}
\usage{
use_data_events(..., overwrite = FALSE)
}
\arguments{
\item{...}{(string) Unquoted name of \code{tibble}/\code{data.frame} to save.}

\item{overwrite}{(logical) By default, \code{use_data_events()} will not
overwrite existing files. If you really want to do so, set this to \code{TRUE}.}
}
\value{
Invisibly returns the location of the saved csv file.
}
\description{
Once data meetings Darwin Core Standard, \code{use_data_events()} makes it
easy to save data in the correct place for building a Darwin Core Archive
with \code{\link[=build_archive]{build_archive()}}.
}
\details{
This function saves data in the \code{data-publish} folder, which is the default
folder that \code{galaxias} uses to build a Darwin Core Archive using
\code{\link[=build_archive]{build_archive()}}.
}
\seealso{
\code{\link[=use_data_occurrences]{use_data_occurrences()}}
}
