% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_data_metadata}
\alias{use_data_metadata}
\title{Add a metadata statement to a Biodiversity Data Package}
\usage{
use_data_metadata()
}
\description{
Add a metadata statement to a Biodiversity Data Package
}
