% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_data_readme}
\alias{use_data_readme}
\title{Add README.Rmd to a Biodiversity Data Package}
\usage{
use_data_readme(package)
}
\description{
This function adds \code{galaxias}-specific \code{README.Rmd} instead of the \code{usethis}
default. This requires replacing placeholder text with the supplied package
name (\code{package}).
}
