% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_datetime.R
\name{use_datetime}
\alias{use_datetime}
\title{Add date and time fields to a \code{tibble}}
\usage{
use_datetime(
  df,
  eventDate = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  time = NULL,
  .keep = "unused"
)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}

\item{eventDate}{The date and/or time that the observation/event occurred.}

\item{year}{The year of the observation/event.}

\item{month}{The month of the observation/event.}

\item{day}{The day of the observation/event.
that contains the whole location, given any possible measurement error.}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{dplyr::mutate}, which defaults to \code{"all"} this defaults to
\code{"unused"}; i.e. only keeps Darwin Core fields, and not those fields used to
generate them.}

\item{eventTime}{The time of the observation/event.}

\item{datePrecision}{Precision of the \code{eventDate}.}
}
\value{
A tibble with the requested fields added.
}
\description{
This function helps format standard date/time fields to a \code{tibble}.
}
\details{
In
practice this is no different from using \code{mutate()}, but gives some
informative errors, and serves as a useful lookup for how spatial fields are
represented in the Darwin Core Standard.

Example values are:
\itemize{
\item \code{geodeticDatum} should be a valid EPSG code
\item \code{coordinatePrecision} should be no less than 0.00001 if data were collected
using GPS
\item \code{coordinateUncertaintyInMeters} will typically be around \code{30} (metres) if
recorded with a GPS after 2000, or \code{100} before that year.
}
}
