% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dwc.R
\name{use_dwc}
\alias{use_dwc}
\alias{as_tibble.dwc_df}
\alias{print.dwc_df}
\title{Convert a \code{tibble} to class \code{dwc_df}}
\usage{
use_dwc(df)

\method{as_tibble}{dwc_df}(x, ...)

\method{print}{dwc_df}(x, ...)
}
\description{
This function is experimental. It tags a tibble with an new class (\code{dwc_df}),
which enables class-specific masked functions from dplyr such as
\code{rename.dwc()} and \code{mutate.dwc()} (relocate? select?).
}
\details{
It appends "dwc_df" as a class ahead of the traditional tibble classes. This
ensures that "dwc_df" methods are called first, but if missing, methods for
class "tibble" still work. This is beneficial as not all functions require
a dwc implementation (e.g. \code{filter()}).
}
