% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_eventDate.R
\name{use_eventDate}
\alias{use_eventDate}
\title{Add \code{event} data to a \code{tibble}}
\usage{
use_eventDate(
  df,
  date_time = NULL,
  date = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  time = NULL,
  .keep = "unused"
)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}

\item{date_time}{an object that combines date and time information}

\item{date}{date only}

\item{year}{year only}

\item{month}{month only}

\item{day}{day only}

\item{time}{time of day}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{dplyr::mutate}, which defaults to \code{"all"} this defaults to
\code{"unused"}; i.e. only keeps Darwin Core fields, and not those fields used to
generate them.}
}
\value{
A tibble with the requested fields.
}
\description{
\code{eventDate} is the main variable here, and describes when an event occurred.
This function takes the information provided, and returns the most specific
information it can on when an event occurred.
}
