% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_occurrences.R
\name{use_id_random}
\alias{use_id_random}
\title{Create a random identifier column}
\usage{
use_id_random(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Uses \code{uuid::UUIDgenerate()} to create a random UUID code without the possible
shortfalls of being influenced by R's internal random number generators
(i.e., set.seed).
}
