% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_observer.R
\name{use_observer}
\alias{use_observer}
\title{Add who made an observation to a \code{tibble}}
\usage{
use_observer(.df, recordedBy = NULL, recordedByID = NULL, .keep = "unused")
}
\arguments{
\item{recordedBy}{Names of people, groups, or organizations responsible for
recording the original occurrence. The primary collector or observer should
be listed first.}

\item{recordedByID}{The globally unique identifier for the person, people,
groups, or organizations responsible for recording the original occurrence.}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{dplyr::mutate}, which defaults to \code{"all"} this defaults to
\code{"unused"}; i.e. only keeps Darwin Core fields, and not those fields used to
generate them.}

\item{df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}
}
\value{
A tibble with the requested fields added.
}
\description{
Format fields that contain information about who made a specific observation
of an organism.
}
\details{
In practice this is no different from using \code{mutate()}, but gives some
informative errors, and serves as a useful lookup for fields in
the Darwin Core Standard.

Examples of \code{recordedBy} values:
\itemize{
\item \verb{José E. Crespo}
}

Examples of \code{recordedByID} values:
\itemize{
\item \code{c("https://orcid.org/0000-0002-1825-0097", "https://orcid.org/0000-0002-1825-0098")}
}
}
