% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_occurrenceID.R
\name{use_occurrenceID}
\alias{use_occurrenceID}
\title{Add an \code{occurrenceID} field to a \code{tibble}}
\usage{
use_occurrenceID(.df, ...)
}
\arguments{
\item{...}{Optional fields to use as a unique identifier. If multiple are
provided, will be concatenated into a single unique identifier.}

\item{df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}
}
\value{
A \code{tibble} with an attached \code{basisOfRecord} field
}
\description{
Every occurrence should have an occurrenceID entry. Ideally these should be
persistent to avoid being lost in future updates. They should also be
unique, both within the dataset, and (ideally) across all other datasets.
}
