% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_occurrences.R
\name{use_occurrences}
\alias{use_occurrences}
\title{Add occurrence-specific information to a \code{tibble}}
\usage{
use_occurrences(
  .df,
  occurrenceID = NULL,
  basisOfRecord = NULL,
  occurrenceStatus = NULL,
  .keep = "unused",
  .messages = TRUE
)
}
\arguments{
\item{occurrenceID}{A character string. Every occurrence should have an
\code{occurrenceID} entry. Ideally IDs should be persistent to avoid being lost
in future updates. They should also be unique, both within the dataset, and
(ideally) across all other datasets.}

\item{basisOfRecord}{Record type. Only accepts \code{camelCase}, for
consistency with field names.
Accepted \code{basisOfRecord} values are one of:
\itemize{
\item \code{"humanObservation"}, \code{"machineObservation"}, \code{"livingSpecimen"},
\code{"preservedSpecimen"}, \code{"fossilSpecimen"}, \code{"materialCitation"}
}}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{dplyr::mutate}, which defaults to \code{"all"} this defaults to
\code{"unused"}, which only keeps Darwin Core fields and not those fields used to
generate them.}

\item{df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}
}
\value{
A tibble with the requested fields added.
}
\description{
Format fields uniquely identify each occurrence record and specify the type
of record. \code{occurrenceID} and \code{basisOfRecord} are necessary fields of
information for occurrence records, and should be appended to a data set
to conform to Darwin Core Standard prior to submission.

In practice this is no different from using \code{mutate()}, but gives some
informative errors, and serves as a useful lookup for fields in
the Darwin Core Standard.
}
\details{
Examples of \code{occurrenceID} values:
\itemize{
\item \verb{000866d2-c177-4648-a200-ead4007051b9}
\item \verb{http://arctos.database.museum/guid/MSB:Mamm:233627}
}

Accepted \code{basisOfRecord} values are one of:
\itemize{
\item \code{"humanObservation"}, \code{"machineObservation"}, \code{"livingSpecimen"},
\code{"preservedSpecimen"}, \code{"fossilSpecimen"}, \code{"materialCitation"}
}
}
