% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_scientific_name.R
\name{use_scientific_name}
\alias{use_scientific_name}
\title{Add scientific name and authorship to a \code{tibble}}
\usage{
use_scientific_name(
  .df,
  scientificName = NULL,
  scientificNameRank = NULL,
  scientificNameAuthorship = NULL,
  .keep = "unused"
)
}
\arguments{
\item{scientificName}{The full scientific name in the lower level taxonomic
rank that can be determined.}

\item{scientificNameRank}{The taxonomic rank of \code{scientificName}.}

\item{scientificNameAuthorship}{The authorship information for \code{scientificName}.}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{dplyr::mutate}, which defaults to \code{"all"} this defaults to
\code{"unused"}; i.e. only keeps Darwin Core fields, and not those fields used to
generate them.}

\item{df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}
}
\value{
A tibble with the requested fields added.
}
\description{
Format the field \code{scientificName}, the lowest identified taxonomic name of an
occurrence, along with the rank and authorship of the provided name.
}
\details{
Examples of \code{scientificName} values (we specify the rank in parentheses, but
users should not include this information):
\itemize{
\item \code{Coleoptera} (order)
\item \code{Vespertilionidae} (family)
\item \code{Manis} (genus)
\item \verb{Ctenomys sociabilis} (genus + specificEpithet)
\item \verb{Ambystoma tigrinum diaboli} (genus + specificEpithet + infraspecificEpithet)
}

Examples of \code{scientificNameRank}:
\itemize{
\item \code{order}
\item \code{genus}
\item \code{subspecies}
\item \code{infraspecies}
}

Examples of \code{scientificNameAuthorship}:
\itemize{
\item \verb{(Györfi, 1952)}
\item \verb{R. A. Graham}
\item \verb{(Martinovský) Tzvelev}
}
}
\seealso{
\code{\link[=use_taxonomy]{use_taxonomy()}} for taxonomic name information.
}
