% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sf.R
\name{use_sf}
\alias{use_sf}
\title{Add spatial fields to a \code{tibble} using \code{sf} \code{POINT} coordinates}
\usage{
use_sf(.df, geometry = NULL, .keep = "unused")
}
\arguments{
\item{.df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}

\item{geometry}{the latitude/longitude coordinates in decimal degrees as \code{sf} \code{POINT} class}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{dplyr::mutate}, which defaults to \code{"all"} this defaults to
\code{"unused"}; i.e. only keeps Darwin Core fields, and not those fields used to
generate them.}
}
\value{
A tibble with the requested fields added.
}
\description{
This function helps format standard location fields to a \code{tibble}. It differs
from \code{use_coordinates()} by accepting \code{sf} geometry columns of class \code{POINT}
as coordinates (rather than \code{numeric} lat/lon coordinates). The advantage
of using an \code{sf} geometry is that the Coordinate Reference System (CRS) is
automatically formatted into the required \code{geodeticDatum} column.
}
\seealso{
\code{\link[=use_coordinates]{use_coordinates()}} for providing numeric coordinates,
\code{\link[=use_locality]{use_locality()}} for providing text-based spatial information
}
