% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_taxonomy.R
\name{use_taxonomy}
\alias{use_taxonomy}
\title{Add taxonomic information to a \code{tibble}}
\usage{
use_taxonomy(
  .df,
  kingdom = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL,
  genus = NULL,
  species = NULL,
  specificEpithet = NULL,
  vernacularName = NULL,
  .keep = "unused"
)
}
\arguments{
\item{kingdom}{The kingdom name of identified taxon.}

\item{phylum}{The phylum name of identified taxon.}

\item{class}{The class name of identified taxon.}

\item{order}{The order name of identified taxon.}

\item{family}{The family name of identified taxon.}

\item{genus}{The genus name of the identified taxon.}

\item{species}{The species name of the identified taxon.}

\item{specificEpithet}{The name of the first species or species epithet of
the \code{scientificName}.
\href{https://dwc.tdwg.org/list/#dwc_specificEpithet}{See documentation}}

\item{vernacularName}{The common or vernacular name of the identified taxon.}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{dplyr::mutate}, which defaults to \code{"all"} this defaults to
\code{"unused"}; i.e. only keeps Darwin Core fields, and not those fields used to
generate them.}

\item{df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}
}
\value{
A tibble with the requested fields added.
}
\description{
Format fields that contain taxonomic name information from kingdom to
species, as well as the common/vernacular name, to a \code{tibble}.
}
\details{
In
practice this is no different from using \code{mutate()}, but gives some
informative errors, and serves as a useful lookup for taxonomic names in
the Darwin Core Standard.

Examples of \code{specificEphithet}:
\itemize{
\item If \code{scientificName} is \verb{Abies concolor}, the \code{specificEpithet} is \code{concolor}.
\item If \code{scientificName} is \verb{Semisulcospira gottschei}, the \code{specificEpithet} is \code{gottschei}.
}
}
\seealso{
\code{\link[=use_scientific_name]{use_scientific_name()}} for adding \code{scientificName} and authorship information.
}
