% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{validate_decimal_latitude}
\alias{validate_decimal_longitude}
\alias{validate_event_date}
\alias{validate_month_day}
\title{Validate
Wrapper for validation checks on data. Logs are saved.
Todo: add something to store results e.g. store <- list()  append true/false
or other results from each validation step for final report output}
\usage{
validate(data)

validate_decimal_latitude(data)

validate_decimal_longitude(data)

validate_event_date(data)

validate_month_day(date)
}
\arguments{
\item{data}{data frame}
}
\value{
Console output

A message to the console

A message to the console

A message to the console

A message to the console
}
\description{
Validate
Wrapper for validation checks on data. Logs are saved.
Todo: add something to store results e.g. store <- list()  append true/false
or other results from each validation step for final report output

Validate latitude
Check that values are within the valid range of -90 to 90

Validate longitude
Check that values are within the valid range of -180 to 180

Validate eventDate
\itemize{
\item validate against accepted regex patterns
\item validate year, month, and day if present
\item see term details \href{https://dwc.tdwg.org/terms/#dwc:eventDate}{here}
}

Validate eventDate
\itemize{
\item validate against some accepted patterns (gbif)
\item detect yyyy-mm-dd formats and validate them
}
}
\keyword{internal}
