% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_archive.R
\name{validate_archive}
\alias{validate_archive}
\title{Validate a Darwin Core Archive via API}
\usage{
validate_archive(directory = ".", filename = NULL, provider = "GBIF")
}
\arguments{
\item{directory}{Path to a nominated archive.}

\item{filename}{Optionally specify a pre-built DwCA instead of a directory.}

\item{provider}{(string) The institution to be queried for validation
services. Currently only \code{"GBIF"} is supported.}
}
\value{
A report on the supplied archive.
}
\description{
Note: Unclear whether this is available yet, from GBIF or ALA.
}
