% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_dwca.R
\name{validate_dwca}
\alias{validate_dwca}
\title{Validate a Darwin Core Archive via API}
\usage{
validate_dwca(pkg = ".", file = NULL, provider = "GBIF")
}
\arguments{
\item{file}{optionally specify a pre-built DwCA instead of a directory}

\item{provider}{(string) the institution to be queried for validation
services. Currently only \code{"GBIF"} is supported.}

\item{path}{path to a nominated Biodiversity Data Package}
}
\value{
A report on the supplied archive.
}
\description{
Validate a Darwin Core Archive via API
}
