% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_eml.R
\name{validate_output_path}
\alias{validate_output_path}
\title{Validate a user input output path / file name}
\usage{
validate_output_path(filepath)
}
\arguments{
\item{filepath}{string Path to output file}
}
\value{
Console report. Stops if validation fails.
}
\description{
Validate a user input output path / file name
}
\keyword{internal}
