% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_table.R
\name{word_overlap_score}
\alias{word_overlap_score}
\title{Word overlap score}
\usage{
word_overlap_score(string1, string2)
}
\arguments{
\item{string1}{string}

\item{string2}{string}
}
\value{
numeric The number of word components that overlap/match between
words (0 no matches, 1 = one match, etc.)
}
\description{
Word overlap score
}
\seealso{
Called by \code{\link[=map_fields]{map_fields()}}; calls \code{\link[=split_into_words]{split_into_words()}}
}
\keyword{internal}
