% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_md.R
\name{write_md}
\alias{write_md}
\title{Write from various formats to Markdown}
\usage{
write_md(x, file, title = "Title goes here")
}
\arguments{
\item{file}{(string) file name to save text, typically to \code{.md} or \code{.Rmd}}

\item{xml}{A xml object, either imported using \code{xml2}, or the url or a valid
xml page.}
}
\description{
Export format is chosen based on the file suffix; GitHub
(.md), Rmarkdown (.Rmd) and Quarto (.qmd) are supported.
}
