% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_rmd.R
\name{xml_to_rmd}
\alias{xml_to_rmd}
\title{wrapper function to convert xml to a .rmd file}
\usage{
xml_to_rmd(xml, file)
}
\arguments{
\item{xml}{An \code{xml} object imported using \code{xml2}}

\item{file}{(string) file name to save text, typically to \code{.md} or \code{.Rmd}}
}
\description{
wrapper function to convert xml to a .rmd file
}
