% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{add_}
\alias{add_}
\alias{add_occurrences}
\alias{add_events}
\alias{add_multimedia}
\alias{add_metadata}
\alias{add_data}
\alias{add_schema}
\title{Add information to a \code{dwca} object}
\usage{
add_occurrences(.dwca, df = NULL)

add_events(.dwca, df = NULL)

add_multimedia(.dwca, df = NULL)

add_metadata(.dwca, xml = NULL)

add_data(.dwca, df = NULL, slot = NULL)

add_schema(.dwca)
}
\arguments{
\item{.dwca}{An object of class \code{dwca}, constructed with \code{dwca()}.}

\item{df}{Object of class \code{tibble} or \code{data.frame}.}

\item{xml}{Object of class \code{xml_document}, such as imported from
\code{.xml} with \code{xml2::read_xml()} or from markdown (\code{.rmd}) using
\code{galaxias::read_md()}.}

\item{slot}{Name of the slot where this data should go.}
}
\value{
An object of class \code{dwca}, updated with the provided information.
}
\description{
These are all pipable functions for adding relevant information to a \code{dwca}
object. The underlying workhorse function is \code{add_data()}, but it is
recommended to use the more specific functions \code{add_occurrences()},
\code{add_events()} and \code{add_media()}. \code{add_metadata()} is also mandatory and
requires either an \code{xml} object/file or a markdown file (\code{rmd} or \code{md}
preferred). \code{add_schema()} is called internally and there is no
requirement for the user to call it, but it is shipped here as it may be
useful for debugging purposes.
}
