% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build a Darwin Core Archive from a specified directory}
\usage{
build(directory = ".")
}
\description{
This function is similar to \code{write_dwca()}, but takes a directory as an
argument, rather than an object of class \code{darwin_core_archive} from the
users workspace. This allows users to build an archive without calling
\code{galaxias} within a piped workflow; but as a result, also performs less
checking on the supplied objects
}
