% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dwca.R
\name{build_dwca}
\alias{build_dwca}
\title{Build a Darwin Core Archive from a Biodiversity Data Package}
\usage{
build_dwca(project = ".", path = NULL)
}
\arguments{
\item{project}{A directory containing a Biodiversity Data Project. Defaults
to current working directory.}

\item{path}{(Optional) Path to the resulting zip file. Defaults to \code{NULL},
indicating that the file should be saved to the current parent directory.}
}
\value{
Called exclusively for the side-effect of building a 'Darwin Core
Archive' (i.e. a zip file); doesn't return anything to the workspace.
}
\description{
Similar to \code{devtools::build()}, but to construct a Darwin Core Archive
rather than an R package.
}
\details{
This function looks for specific objects in specific places, namely:
\itemize{
\item Named objects in the \code{data} folder
\item \code{metadata.md} in base folder (or \code{inst} folder as a backup)
}

If either data or metadata are missing missing, the function will
terminate with an error.

Note that Darwin Core Archives also require a 'schema' file. This is
buily internally using \code{build_schema()}.
}
