% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dwca.R
\name{build_dwca}
\alias{build_dwca}
\title{Build a Darwin Core Archive from a \code{dwca} object}
\usage{
build_dwca(.dwca, file = "dwca.zip")
}
\arguments{
\item{.dwca}{A \code{dwca} object}

\item{path}{Name of the zip file}
}
\value{
No object is returned; this function is called for the side-effect
of building a 'Darwin Core Archive' (i.e. a zip file)
}
\description{
Build a Darwin Core Archive from a \code{dwca} object
}
