% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata.R
\name{build_metadata}
\alias{build_metadata}
\title{Create a metadata statement for a Darwin Core Archive}
\usage{
build_metadata(source = "metadata.md", destination = "./data/eml.xml")
}
\arguments{
\item{source}{A metadata file stored in markdown format (\code{.md}). Defaults
to \code{metadata.md}, which is the same as is created by \code{\link[=use_metdata]{use_metdata()}}}

\item{destination}{A file where the result should be saved. Defaults to
\code{data/eml.xml}.}
}
\value{
Does not return an object to the workspace; called for the side
effect of building a file named \code{meta.xml} in the \code{data} directory.
}
\description{
A metadata statement lists the owner of the dataset, how it was collected,
and how it can be used (i.e. its' licence). This function simply reads
converts metadata stored in a markdown file, converts it to xml, and saves it
in the \code{destination} file.
}
\details{
This function is a fairly shallow wrapper on top of functionality build
in the \code{delma} package, particularly \code{\link[delma:read_md]{delma::read_md()}} and
\code{\link[delma:write_eml]{delma::write_eml()}}. You can use that package to gain greater control, or to
debug problems, should you wish.
}
