% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata.R
\name{build_metadata}
\alias{build_metadata}
\title{Create a metadata statement for a Darwin Core Archive}
\usage{
build_metadata(file)
}
\arguments{
\item{file}{path to a metadata statement stored in markdown format (.md).}
}
\value{
Does not return an object to the workspace; called for the side
effect of building a file named \code{meta.xml} in the \code{data} directory.
}
\description{
A metadata statement lists the owner of the dataset, how it was collected,
and how it may used (i.e. its' licence). This function simply converts
metadata stored in a markdown file to xml, and stores it in the \code{data}
folder.
}
\details{
In practice, this is a fairly shallow wrapper on top of functionality build
in the \code{elm} package; you can use that package to gain greater control
should you wish.
}
