% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_schema.R
\name{build_schema}
\alias{build_schema}
\title{Create meta.xml file}
\usage{
build_schema(data, file_name = "occurrence.csv")
}
\arguments{
\item{data}{data frame Validated DwC dataset}

\item{file_name}{file name of DwC standard occurrence data}
}
\value{
xml document
}
\description{
A metafile is required when an archive includes any extension files or if a
}
\examples{
DwC_occurrence_data <-
  tibble(
    basisOfRecord = "HUMAN_OBSERVATION",
    scientificName = "Galaxias maculatus",
    eventDate = "2002-12-07",
    decimalLatitude = -33.366551,
    decimalLongtitude = 151.47635
  )

make_core_xml(DwC_occurrence_data)
}
