% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_schema.R
\name{build_schema}
\alias{build_schema}
\title{Create a \code{schema} for a Darwin Core Archive}
\usage{
build_schema(directory = ".")
}
\arguments{
\item{project}{a directory containing Darwin Core data, preferably built
with \code{use_bd_project()}.}
}
\value{
Does not return an object to the workspace; called for the side
effect of building a file named \code{meta.xml} in the specified directory.
}
\description{
A schema is an xml document that maps the files and field names in a DwCA.
This function is intended to be primarily internal, and is called by
\code{build_dwca()}, but is provided for debugging purposes. It works on csv files
in a specified directory.
}
