% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_archive.R
\name{check_archive}
\alias{check_archive}
\title{Check an archive against Darwin Core standards}
\usage{
check_archive(source = "data-publish")
}
\arguments{
\item{source}{(string) A directory containing the files to be published, or
optionally a \code{.zip} file built from the same (i.e. with \code{build_archive()}).
Defaults to the \code{data-publish} folder within the current working directory.}
}
\value{
Invisibly returns a tibble to the workspace containing check
results; but primarily called for the side-effect of generating a report in
the console.
}
\description{
Check whether schema and EML files (i.e. \code{meta.xml} and \code{eml.xml}) are
valid with the \code{delma} package; csv files are checked with the \code{corella}
package. This is a wrapper to two other packages.
}
\seealso{
\code{validate_archive()} which runs checks via API, rather than locally.
}
