% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_archive.R, R/get_report.R
\name{check_archive}
\alias{check_archive}
\alias{get_report}
\title{Check whether a Darwin Core Archive meets Darwin Core Standard via API}
\usage{
check_archive(source = "dwc-archive.zip", provider = "GBIF")

get_report(key)
}
\arguments{
\item{source}{(string) A file name (ending in .zip) of the specified Darwin
Core Archive.
Defaults to the \code{dwc-archive.zip} within the top folder of the current
working directory.}

\item{provider}{(string) The institution to be queried for validation
services. Currently only \code{"GBIF"} is supported.}

\item{key}{A unique identifier key from GBIF}
}
\value{
Invisibly returns a tibble to the workspace containing validation
results; but primarily called for the side-effect of generating a report in
the console.
}
\description{
Check whether a specified Darwin Core Archive is ready for
sharing and publication, according to the Darwin Core Standard.
\code{check_archive()} tests the specified archive using an online validation
service by sending the archive via an API and returning the results.
Currently only supports validation using GBIF.
}
\details{
Results returned by \code{check_archive()} are directly from the institution API,
not from galaxias.
}
\examples{
\dontrun{
# add GBIF login details
galaxias_config(gbif = list(username = "your-gbif-username",
                            email = "your-gbif-email",
                            password = "your-gbif-password"))

# Check archive against Darwin Core Standard criteria
check_archive("dwc-archive.zip")
}
}
\seealso{
\code{check_directory()} which runs checks on a folder directory locally,
rather than via API.
}
