% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_directory.R
\name{check_directory}
\alias{check_directory}
\title{Check whether contents of directory comply with the Darwin Core Standard}
\usage{
check_directory()
}
\value{
Doesn't return anything; called for the side-effect of generating a
report in the console.
}
\description{
Checks that files in the \code{data-publish} directory meet Darwin Core Standard.
\code{check_directory()} runs \code{\link[corella:check_dataset]{corella::check_dataset()}} on \code{occurrences.csv} and
\code{events.csv} files, and \code{\link[delma:check_metadata]{delma::check_metadata()}} on the \code{eml.xml}
file, if they are present. These \code{check_} functions run tests to determine
whether data and metadata pass Darwin Core Standard criteria.
}
\seealso{
\code{\link[=check_archive]{check_archive()}} checks a Darwin Core Archive via a GBIF API,
rather than locally.
}
