% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dwc.R, R/use_occurrenceID.R,
%   R/use_occurrences.R, R/use_locality.R, R/use_coordinates.R,
%   R/use_coordinates_sf.R, R/use_datetime.R
\name{check_dwc}
\alias{check_dwc}
\alias{check_fields}
\alias{check_occurrenceID}
\alias{check_basisOfRecord}
\alias{check_continent}
\alias{check_country}
\alias{check_countryCode}
\alias{check_decimalLatitude}
\alias{check_coords}
\alias{check_is_sf}
\alias{col_sf_to_dwc}
\alias{check_eventDate}
\alias{check_time}
\alias{check_year}
\alias{check_month}
\alias{check_day}
\alias{check_decimalLongitude}
\alias{check_geodeticDatum}
\alias{check_crs}
\title{Check for Darwin Core field conformance}
\usage{
check_dwc(.df)

check_fields(.df, level = c("inform", "warn", "abort"))

check_occurrenceID(.df, level = c("inform", "warn", "abort"))

check_basisOfRecord(.df, level = c("inform", "warn", "abort"))

check_continent(df, level = c("inform", "warn", "abort"))

check_country(df, level = c("inform", "warn", "abort"))

check_countryCode(df, level = c("inform", "warn", "abort"))

check_decimalLatitude(.df, level = c("inform", "warn", "abort"))

check_coords(.df, level = c("inform", "warn", "abort"))

check_is_sf(.df, level = c("inform", "warn", "abort"), call = caller_env())

col_sf_to_dwc(.df, level = c("inform", "warn", "abort"), call = caller_env())

check_eventDate(.df, level = c("inform", "warn", "abort"))

check_time(.df, level = c("inform", "warn", "abort"))

check_year(.df, level = c("inform", "warn", "abort"))

check_month(.df, level = c("inform", "warn", "abort"))

check_day(.df, level = c("inform", "warn", "abort"))

check_decimalLongitude(.df, level = c("inform", "warn", "abort"))

check_geodeticDatum(.df, level = c("inform", "warn", "abort"))

check_crs(.df, level = "warn", call = caller_env())
}
\arguments{
\item{level}{what action should the function take for non-conformance?
Defaults to \code{"inform"}.}

\item{df}{a tibble containing data}
}
\description{
Function to check whether a data.frame or tibble conforms to DwC standards
NOTE: Option to use \code{bdc} for checks, and/or \code{pointblank} for running them
}
\keyword{Internal}
