% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dwc.R, R/use_occurrenceID.R,
%   R/use_basisOfRecord.R, R/use_locality.R, R/use_coordinates.R,
%   R/use_datetime.R
\name{check_dwc}
\alias{check_dwc}
\alias{check_fields}
\alias{check_occurrenceID}
\alias{check_basisOfRecord}
\alias{check_continent}
\alias{check_country}
\alias{check_countryCode}
\alias{check_decimalLatitude}
\alias{check_eventDate}
\alias{check_year}
\alias{check_eventTime}
\alias{check_decimalLongitude}
\title{Check for Darwin Core field conformance}
\usage{
check_dwc(.df)

check_fields(.df, level = c("inform", "warn", "abort"))

check_occurrenceID(.df, level = c("inform", "warn", "abort"))

check_basisOfRecord(.df, level = c("inform", "warn", "abort"))

check_continent(df, level = c("inform", "warn", "abort"))

check_country(df, level = c("inform", "warn", "abort"))

check_countryCode(df, level = c("inform", "warn", "abort"))

check_decimalLatitude(.df, level = c("inform", "warn", "abort"))

check_eventDate(.df, level = c("inform", "warn", "abort"))

check_year(.df, level = c("inform", "warn", "abort"))

check_eventTime(.df, level = c("inform", "warn", "abort"))

check_decimalLongitude(.df, level = c("inform", "warn", "abort"))
}
\arguments{
\item{level}{what action should the function take for non-conformance?
Defaults to \code{"inform"}.}

\item{df}{a tibble containing data}
}
\description{
Function to check whether a data.frame or tibble conforms to DwC standards
NOTE: Option to use \code{bdc} for checks, and/or \code{pointblank} for running them
}
