% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_metadata}
\alias{check_metadata}
\title{Check that a supplied metadata statement is valid according to GBIF schema}
\usage{
check_metadata(metadata)
}
\arguments{
\item{metadata}{xml object representing metadata}
}
\value{
A tibble showing parsed errors
}
\description{
Note: this function requires the user to be online.
}
