% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bd_package.R
\name{create_bd_package}
\alias{create_bd_package}
\title{Create a new Biodiversity Data Package}
\usage{
create_bd_package(path, copyright_holder = NULL)
}
\arguments{
\item{path}{A path to the directory to create}
}
\description{
Sets up a 'Biodiversity Data Package', which is a modified R package for
storing biodiversity data, ideally according to the Darwin Core Standard.
Based heavily on usethis::create_tidy_package
}
