% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.R
\name{detect_required_metadata}
\alias{detect_required_metadata}
\title{Detect required fields in metadata template}
\usage{
detect_required_metadata(metadata_tibble)
}
\arguments{
\item{metadata_tibble}{}
}
\description{
Detect required fields in metadata template
}
