% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.R
\name{edit_metadata_md}
\alias{edit_metadata_md}
\title{Edit metadata template}
\usage{
edit_metadata_md(path = ".")
}
\arguments{
\item{path}{file path where metadata template will be copied to}
}
\value{
template file opened in RStudio as
}
\description{
Edit metadata template
}
\examples{
edit_metadata_md()
}
