% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{galaxias-dplyr}
\alias{galaxias-dplyr}
\alias{rename.dwc_df}
\title{Functions masked from dplyr}
\usage{
\method{rename}{dwc_df}(.data, ...)
}
\description{
These functions are masked for class \code{dwc} to support type-specific data
checks. Experimental
}
