% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxias-package.R
\docType{package}
\name{galaxias-package}
\alias{galaxias}
\alias{galaxias-package}
\title{Tools for building and sharing Biodiversity Data Packages}
\description{
\code{galaxias} provides tools to build Biodiversity Data Packages, which are R
packages designed to store, document, and share biodiversity information.
Biodiversity Data Packages are fully compatible with other elements of the R
package workflow, including \code{devtools} for building, \code{usethis} for adding
package elements, \code{pkgdown} for converting package contents into webpages.

Once a package is built, \code{galaxias} also provides tools to assess
compatibility of the data with the 'Darwin Core' standard, convert the
package into a valid 'Darwin Core Archive', and publish that archive to a
data infrastructure (though currently only the Atlas of Living Australia is
formally supported).

\code{galaxias} is named for a genus of Australian freshwater fish.
}
\section{\code{usethis} extensions}{


These functions can be identified as being structurally similar to \code{usethis}
functions, but with an added \verb{_data_} infix.
\itemize{
\item \code{\link[=create_data_package]{create_data_package()}} to create a new Biodiversity Data Package
\item \code{\link[=use_data_citation]{use_data_citation()}} to give advice on citing your package
\item \code{\link[=use_data_description]{use_data_description()}} update the description to standard for this package type
\item \code{\link[=use_data_examples]{use_data_examples()}} build \code{data-raw} and \code{data} folders, with example code
\item \code{\link[=use_data_metadata]{use_data_metadata()}} add a metadata statement to \code{vignettes} folder
\item \code{\link[=use_data_readme]{use_data_readme()}} replace the standard README.Rmd with a type-specific version
\item \code{\link[=use_data_schema]{use_data_schema()}} not implemented yet
\item \code{\link[=use_data_tests]{use_data_tests()}} add tests for conformance with Darwin Core standards
}
}

\section{Object classes}{

\code{galaxias} supports the following object classes:
\itemize{
\item \code{dwc_a} Darwin Core Archive objects
\item \code{dwc_df} data frames to be checked against the Darwin Core standard
\item \code{dwc_xml} xml objects to be checked against the Darwin Core standard
}

All can be created using the \code{galaxias()} function
}

\section{Functions}{

\strong{Setting up an archive}
\itemize{
\item \code{\link[=dwca]{dwca()}} create pipeable \code{dwca} objects
\item \code{\link[=add_occurrences]{add_occurrences()}}, \code{\link[=add_events]{add_events()}}, \code{\link[=add_multimedia]{add_multimedia()}}, \code{\link[=add_metadata]{add_metadata()}} to add data to \code{dwca} objects
\item \code{\link[=read_dwca]{read_dwca()}}, \code{\link[=write_dwca]{write_dwca()}}
}

\strong{Data handling}
\itemize{
\item \code{\link[=build_random_identifier]{build_random_identifier()}}, \code{\link[=build_composite_identifier]{build_composite_identifier()}} uuids
\item \code{\link[=get_metadata_template]{get_metadata_template()}}, \code{\link[=get_metadata_example]{get_metadata_example()}} get example metadata in xml
\item \code{\link[=read_md]{read_md()}}, \code{\link[=write_md]{write_md()}} to convert between markdown and xml
}

\strong{Checking and publication}
\itemize{
\item \code{\link[=build]{build()}} build a \code{dwca} into a zip file
\item \code{\link[=test]{test()}} run local tests on an archive
\item \code{\link[=check]{check()}} check a \code{dwca} conforms to archive-level standards
\item \code{\link[=publish]{publish()}} send your data to the ALA for publication
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Fonti Kar \email{fonti.kar@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
  \item Andrew Schwenke \email{andrew.schwenke@csiro.au}
}

}
\keyword{internal}
