% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxias-package.R
\docType{package}
\name{galaxias-package}
\alias{galaxias}
\alias{galaxias-package}
\title{Tools for building and sharing Biodiversity Data Packages}
\description{
\code{galaxias} provides tools to build Biodiversity Data Packages, which are R
packages designed to store, document, and share biodiversity information.
Biodiversity Data Packages are fully compatible with other elements of the R
package workflow, including \code{devtools} for handling your package-buidling
workflow, \code{usethis} for adding package elements, \code{pkgdown} for displaying
package contents as webpages.

Once a package is built, \code{galaxias} also provides tools to assess
compatibility of the data with the 'Darwin Core' standard, convert the
package into a valid 'Darwin Core Archive', and publish that archive to a
data infrastructure (though currently only the Atlas of Living Australia is
formally supported).

\code{galaxias} is named for a genus of Australian freshwater fish.
}
\section{Functions}{

\strong{Building your project}

The easiest way to get started is to use \code{\link[=create_bd_package]{create_bd_package()}} to create a
new project with all the required files and folders. Internally,
\code{\link[=create_bd_package]{create_bd_package()}} calls the following functions:
\itemize{
\item \code{\link[=use_bd_description]{use_bd_description()}} update the description to standard for this package type
\item \code{\link[=use_bd_readme_rmd]{use_bd_readme_rmd()}} Add a README.Rmd file with advice for this type of package
\item \code{\link[=use_bd_data_raw]{use_bd_data_raw()}} build \code{data-raw} folder, with example code
}

Note the syntax here is based on\code{ usethis}. The next step is to add data
to the \code{data-raw} folder and process it. Once that is complete, the following
functions are required:
\itemize{
\item \code{\link[=use_bd_data]{use_bd_data()}} to create the requisite 'processed' data files
\item \code{\link[=use_bd_metadata]{use_bd_metadata()}} to create a metadata statement using \code{DESCRIPTION} as a basis
\item \code{\link[=use_bd_schema]{use_bd_schema()}} to create a schema based on objects in the \code{data} folder
}

Finally, the following functions are optional:
\itemize{
\item \code{\link[=use_bd_citation]{use_bd_citation()}} to give advice on citing your package
\item \code{\link[=use_bd_testthat]{use_bd_testthat()}} add tests for conformance with Darwin Core standards
\item \code{\link[=use_bd_vignette]{use_bd_vignette()}} add a vignette that reports on basic stats of your package
}

\strong{Data manipulation functions}

These functions assist with populating specific Darwin Core fields.
Q: should these have \code{use_} prefix? \code{add_}?
NOTE: should integrate properly with \code{mutate()} and friends.
\itemize{
\item \code{\link[=use_occurrenceID]{use_occurrenceID()}} add a unique identifier, either using random numbers or a composite of existing columns
\item \code{\link[=use_basisOfRecord]{use_basisOfRecord()}} add a basisOfRecord field
\item use_eventDate
\item use_decimalLatitude()
\item use_decimalLongtiude()
\item etc.
\item \code{\link[=dwc_fields]{dwc_fields()}} NOT IMPLEMENTED YET, but would be useful e.g. \code{df |> select(any_of(dwc_fields()))}
}

\strong{testthat integration}

Again, syntax and utility not locked down yet. Example code in \code{testthat.R}
\itemize{
\item \code{\link[=test_metadata]{test_metadata()}} check xml metadata is valid using GBIF (or just call in \code{vignettes}?)
\item \code{\link[=skip_dwc]{skip_dwc()}} report only on DwC objects and fields that are actually present
}

\strong{Checking and publication}
\itemize{
\item \code{\link[=build_dwca]{build_dwca()}} convert a Biodiversity Data package to a Darwin Core Archive (DwCA)
\item \code{\link[=validate_dwca]{validate_dwca()}} check your archive via the ALA validate API (NOT FUNCTIONAL)
\item \code{\link[=publish_dwca]{publish_dwca()}} send your data to the ALA for publication (NOT FUNCTIONAL)
}

\strong{Internal Functions}
\itemize{
\item \code{\link[=read_md]{read_md()}} read a markdown file into R as xml
\item \code{\link[=write_md]{write_md()}} convert an xml document in the workspace to a markdown file
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
}

}
\keyword{internal}
