% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxias-package.R
\docType{package}
\name{galaxias-package}
\alias{galaxias}
\alias{galaxias-package}
\title{Build repositories to share biodiversity data}
\description{
\code{{galaxias}} is a \code{{usethis}} extension for building 'Darwin Core Archives'; a
global data standard to store, document, and share biodiversity information.
The package provides tools build data projects, manipulate data to
conform with the standard, and check validity against established norms.

The package is named for a genus of freshwater fish.
}
\section{Functions}{

\strong{Building your project}

The default method to start your project is to call \code{\link[=create_bd_project]{create_bd_project()}}.
Alternatively, you can call the following functions to add specific content
to an existing directory:
\itemize{
\item \code{\link[=use_bd_data]{use_bd_data()}} to create the requisite 'processed' data files
\item \code{\link[=use_bd_data_raw]{use_bd_data_raw()}} build \code{data-raw} folder, with example code
\item \code{\link[=use_bd_description]{use_bd_description()}} add a \code{DESCIRPTION} file
\item \code{\link[=use_bd_metadata]{use_bd_metadata()}} to create a metadata statement in the home directory
\item \code{\link[=use_bd_readme_md]{use_bd_readme_md()}}, \code{\link[=use_bd_readme_rmd]{use_bd_readme_rmd()}} Add a \code{README.Rmd} file with advice for biodiversity data
}

\strong{Add Darwin Core Terms}

The following functions add single DwC fields, or collections of related
fields, to an existing \code{tibble}.
\itemize{
\item \code{\link[=use_coordinates]{use_coordinates()}} for spatial data
\item \code{\link[=use_eventDate]{use_eventDate()}} for temporal data
\item \code{\link[=use_locality]{use_locality()}} for spatial descriptions
}

Proposed:
\itemize{
\item \code{\link[=use_occurrences]{use_occurrences()}} basic information on observations (occurrenceID, basisOfrecord, recordID (?))
\item \code{\link[=use_events]{use_events()}} basic information on observation events (eventID, parentEventID)
\item \code{\link[=use_scientificName]{use_scientificName()}} record the highest level of taxonomic specificity in the dataset (scientificName, scientificNameRank, scientificNameAuthorship)
\item \code{\link[=use_taxonomy]{use_taxonomy()}} to specify higher taxonomic columns (kingdom, phylum, class, order, family, genus, species, specificEpithet, vernacularName)
\item \code{\link[=use_abundance]{use_abundance()}} to state how many animals were seen during the observation (individualCount, organismQuantity, organismQuantityType, occurrenceStatus)
}

Possible functions for added functionality:
\itemize{
\item \code{\link[=use_darwin_core]{use_darwin_core()}} to subset to only fields with DwC names (i.e. same as \code{df |> select(any_of(dwc_fields()))})
\item \code{\link[=use_individuals]{use_individuals()}} attributes of individuals measured (individualID, lifeStage, sex, vitality, reproductiveCondition)
\item \code{\link[=use_observer]{use_observer()}} to specify who made the observation (recordedByID, recordedBy)
\item \code{\link[=use_collection]{use_collection()}} to give museum- or collection- specific information (datasetID, datasetName, catalogNumber)
\item \code{\link[=use_measurement]{use_measurement()}} for 'Measurement or Fact' data (optional rn)
\item \code{\link[=use_media]{use_media()}} good idea, but unclear how users would supply said media; should be urls, but to where?
}

\strong{Checking data for Darwin Core compliance}

The wrapper function for checking tibbles for Darwin Core compliance is
\code{\link[=check_dwc]{check_dwc()}}. It calls the following microfunctions:
\itemize{
\item \code{\link[=check_fields]{check_fields()}} Checks whether non-DwC fields are presen
\item \code{\link[=check_occurrenceID]{check_occurrenceID()}}
\item \code{\link[=check_basisOfRecord]{check_basisOfRecord()}}
\item \code{\link[=check_continent]{check_continent()}}
\item \code{\link[=check_country]{check_country()}}
\item \code{\link[=check_countryCode]{check_countryCode()}}
\item \code{\link[=check_decimalLatitude]{check_decimalLatitude()}}
\item \code{\link[=check_decimalLongitude]{check_decimalLongitude()}}
}

Note that there are more \code{check_} functions than \code{use_} functions, because
some \code{use_} functions affect multiple fields.

\strong{DwCA functions}
\itemize{
\item \code{\link[=build_schema]{build_schema()}} Identify DwC csvs in a directory and build a schema file (\code{meta.xml})
\item \code{\link[=build_metadata]{build_metadata()}} Supply a metadata file in .md format, and convert to a metadata file (\code{eml.xml})
\item \code{\link[=build_dwca]{build_dwca()}} Convert a Biodiversity Data directory to a Darwin Core Archive (DwCA)
\item \code{\link[=validate_dwca]{validate_dwca()}} check your archive via the ALA validate API
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
}

}
\keyword{internal}
