% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxias-package.R
\docType{package}
\name{galaxias-package}
\alias{galaxias}
\alias{galaxias-package}
\title{Tools for building and sharing Biodiversity Data Packages}
\description{
\code{galaxias} provides tools to build Biodiversity Data Packages, which are R
packages designed to store, document, and share biodiversity information.
Biodiversity Data Packages are fully compatible with other elements of the R
package workflow, including \code{devtools} for handling your package-buidling
workflow, \code{usethis} for adding package elements, \code{pkgdown} for displaying
package contents as webpages.

Once a package is built, \code{galaxias} also provides tools to assess
compatibility of the data with the 'Darwin Core' standard, convert the
package into a valid 'Darwin Core Archive', and publish that archive to a
data infrastructure (though currently only the Atlas of Living Australia is
formally supported).

\code{galaxias} is named for a genus of Australian freshwater fish.
}
\section{Functions}{

\strong{Directory & file type manipulation}
\itemize{
\item \code{\link[=create_bd_package]{create_bd_package()}} to create a new Biodiversity Data Package
\item \code{\link[=build_dwca]{build_dwca()}} convert a Biodiversity Data package to a Darwin Core Archive (DwCA) - NOT IMPLEMENTED YET
\item \code{\link[=build_bd_package]{build_bd_package()}} convert a DwCA to a Biodiversity Data Package
\item \code{\link[=read_md]{read_md()}} read a markdown file into R as xml
\item \code{\link[=write_md]{write_md()}} convert an xml document in the workspace to a markdown file
}

\strong{Add content to Biodiversity Data Packages}

These functions are structurally similar to \code{usethis} functions, but with an
added \verb{_bd_} infix (where \code{bd} stands for \verb{biodiversity data}).

The following functions are called internally by \code{create_bd_package()}, in
this order:
\itemize{
\item \code{\link[=use_bd_description]{use_bd_description()}} update the description to standard for this package type
\item \code{\link[=use_bd_readme]{use_bd_readme()}} replace the standard README.Rmd with a type-specific version
\item \code{\link[=use_bd_data_raw]{use_bd_data_raw()}} build \code{data-raw} folder, with example code
}

The following functions are optional:
\itemize{
\item \code{\link[=use_bd_citation]{use_bd_citation()}} to give advice on citing your package
\item \code{\link[=use_bd_testthat]{use_bd_testthat()}} add tests for conformance with Darwin Core standards
}

\strong{Add Darwin Core fields to an existing data.frame}

These functions assist with populating specific Darwin Core fields.
Q: should these have \code{use_} prefix? \code{add_}?
\itemize{
\item \code{\link[=use_occurrenceID]{use_occurrenceID()}} add a unique identifier, either using random numbers or a composite of existing columns
\item \code{\link[=use_basisOfRecord]{use_basisOfRecord()}} add a basisOfRecord field
\item use_eventDate
\item use_decimalLatitude()
\item use_decimalLongtiude()
\item etc.
\item \code{\link[=dwc_fields]{dwc_fields()}} NOT IMPLEMENTED YET, but would be useful e.g. \code{df |> select(any_of(dwc_fields()))}
}

\strong{testthat integration}

Again, syntax and utility not locked down yet. Example code in \code{testthat.R}
\itemize{
\item \code{\link[=test_metadata]{test_metadata()}} check xml metadata is valid using GBIF (or just call in \code{vignettes}?)
\item \code{\link[=skip_dwc]{skip_dwc()}} report only on DwC objects and fields that are actually present
}

\strong{Checking and publication}

NOTE: these APIs do not function (yet). JWT access required.
\itemize{
\item \code{\link[=validate_dwca]{validate_dwca()}} check your archive via the ALA validate API
\item \code{\link[=publish_dwca]{publish_dwca()}} send your data to the ALA for publication
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
}

}
\keyword{internal}
