% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxias-package.R
\docType{package}
\name{galaxias}
\alias{galaxias}
\alias{galaxias-package}
\title{Building Darwin Core Archives for the Atlas of Living Australia}
\description{
The Darwin Core Archive (DwCA) is a transferable unit of
biodiversity data and associated metadata that allows facilities
like the Atlas of Living Australia (ALA) to efficiently integrate and share
openly with the community. Creating DwCAs by hand is a notoriously fiddly,
task. 'galaxias' enables users to efficiently reformat their data to Darwin
Core Standards and generate production-ready DwCA in a reproducible and
transparent workflow.

\code{galaxias} is named for a genus of Australian freshwater fish.
}
\section{Functions}{

\strong{Set up a DWCA}
\itemize{
\item \code{\link[=dwca]{dwca()}} create pipable \code{dwca} objects
\item \code{\link[=add_occurrences]{add_occurrences()}}, \code{\link[=add_events]{add_events()}}, \code{\link[=add_multimedia]{add_multimedia()}}, \code{\link[=add_metadata]{add_metadata()}} to add data to \code{dwca} objects
}

\strong{Data handling}
\itemize{
\item \code{\link[=build_random_identifier]{build_random_identifier()}}, \code{\link[=build_composite_identifier]{build_composite_identifier()}} uuids
}

\strong{Metadata handling}
\itemize{
\item \code{\link[=get_blank_metadata]{get_blank_metadata()}}, \code{\link[=get_example_metadata]{get_example_metadata()}} get example metadata in xml
\item \code{\link[=read_md]{read_md()}}, \code{\link[=write_md]{write_md()}} to convert between markdown and xml
}

\strong{Concluding}
\itemize{
\item \code{\link[=check_dwca]{check_dwca()}} check a \code{dwca} object for compliance with DwC
\item \code{\link[=build_dwca]{build_dwca()}} build a \code{dwca} into a zip file
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Fonti Kar \email{fonti.kar@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
  \item Andrew Schwenke \email{andrew.schwenke@csiro.au}
}

}
\keyword{internal}
