% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_galaxias-package.R
\docType{package}
\name{galaxias}
\alias{galaxias}
\alias{galaxias-package}
\title{Building Darwin Core Archives for the Atlas of Living Australia}
\description{
The Darwin Core Archive (DwCA) is a transferable unit of
biodiversity data and associated metadata that allows facilities
like the Atlas of Living Australia (ALA) to efficiently integrate and share
openly with the community. Creating DwCAs by hand is a notoriously fiddly,
task. 'galaxias' enables users to efficiently reformat their data to Darwin
Core Standards and generate production-ready DwCA in a reproducible and
transparent workflow.

\code{galaxias} is named for a genus of Australian freshwater fish.
}
\section{Functions}{

\strong{Prepare data}
\itemize{
\item \code{\link[=tibble_to_dwc]{tibble_to_dwc()}}
}

\strong{Build eml.XML}
\itemize{
\item \code{\link[=edit_metadata_md]{edit_metadata_md()}}
\item \code{\link[=read_metadata_md]{read_metadata_md()}}
\item \code{\link[=md_to_tibble]{md_to_tibble()}}
\item \code{\link[=detect_required_metadata_fields]{detect_required_metadata_fields()}}
\item \code{\link[=md_tibble_to_list]{md_tibble_to_list()}}
\item \code{\link[=update_title]{update_title()}}
\item \code{\link[=update_names]{update_names()}}
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Dax Kellie \email{dax.kellie@csiro.au}

Authors:
\itemize{
  \item Martin Westgate \email{martin.westgate@csiro.au}
  \item Fonti Kar \email{fonti.kar@csiro.au}
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Margot Schneider \email{margot.schneider@csiro.au}
}

}
\keyword{internal}
