% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fields_console.R, R/map_fields_shiny.R,
%   R/map_fields_tibble.R
\name{map_fields}
\alias{map_fields}
\alias{map_fields_console}
\alias{map_fields_shiny}
\alias{map_fields_tibble}
\title{Map User Columns to Darwin Core Terms}
\usage{
map_fields_console(data)

map_fields_shiny(df)

map_fields_tibble(mapping, index, dataset = NULL)
}
\arguments{
\item{data}{data frame Containing a dataset of which to map DwC
terms to.}

\item{mapping}{data frame Containing at least two columns}

\item{index}{numeric vector of length 2. The first element is the column
index to use for the data column names, the second element is the column
index to use for the DwC terms e.g. \code{index = c(1, 2)}.}

\item{dataset}{data frame Optional dataset to which the lookup table
pertains, used for validation.}
}
\value{
data frame With renamed columns based on the mapping. Exiting the
mapping early (by using the \code{exit} command when prompted) will return a data
frame with any mappings that were made prior to exiting.
}
\description{
For now, mappings are used to rename columns. Can be modified or extended to
output a dictionary for the meta.xml and preserve column names etc.

This interactive function assists in mapping user data columns to the
standard Darwin Core (DwC) terms. It suggests potential DwC matches for each
column name in the user's dataset. Suggestions are generated based on word
component matches and string distance metrics to provide the best possible
matches.

Suggestions are presented to the user in a numbered list, and the user is
then prompted to select the correct DwC term. If no correct match was found,
the user is prompted to manually input the correct term if known, or skip the
column, before continuing to the next column.

Note that a minimum of five matches is always returned, even if they are
only low confidence matches.

Imports a mapping table, ensuring dataset column names align with DwC terms.
\itemize{
\item Requires a data frame with at least two columns:
\itemize{
\item First: Column names of your dataset.
\item Second: Corresponding DwC terms.
}
\item Each row must correctly pair a dataset column name with a DwC term, e.g.,
\code{data.frame("sci name", "scientificName")}.
}
}
\section{dev notes}{

In progress, potential use cases:
\itemize{
\item meta xml mapping process
\item E2E use; pipeline w/ existing lookup table
\item Importing a lookup table created in a previous session (with \code{map2dwc()})
}
}

