% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_md.R
\name{read_md}
\alias{read_md}
\title{Read markdown format to a \code{tibble}}
\usage{
read_md(file)
}
\arguments{
\item{file}{A markdown file}
}
\value{
An object of class \code{tbl_df}, \code{tbl} and \code{data.frame}, aka a \code{tibble},
containing the following fields:
\itemize{
\item \code{level} header level
\item \code{label} header label
\item \code{attributes} any html node attributes
\item \code{text} any text below that heading, noting empty headers are not retained.
}
}
\description{
Markdown documents can be parsed to xml if we assume that the header level
(typically shown using '#') reflects the nestedness structure of the
dataset. \code{read_md()} converts markdown to a tibble. If html tags are used in
the document (e.g. \verb{<h1>} etc.), their attributes are retained in the
resulting outputs.
}
\details{
Code blocks are not formally supported, so please use text only!
}
