% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_md.R, R/write_md.R
\name{read_md}
\alias{read_md}
\alias{write_md}
\title{Convert between markdown and xml}
\usage{
read_md(file)

write_md(xml, file, title = "Title goes here")
}
\arguments{
\item{file}{(string) file name to save text, typically to \code{.md} or \code{.Rmd}}

\item{xml}{A xml object, either imported using \code{xml2}, or the url or a valid
xml page.}
}
\description{
With \code{write_md()}, export format is chosen based on the file suffix; GitHub
(.md), Rmarkdown (.Rmd) and Quarto (.qmd) are supported.
}
