% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_md.R
\name{read_metadata_md}
\alias{read_metadata_md}
\title{Read metadata md information into R}
\usage{
read_metadata_md(path)
}
\arguments{
\item{path}{file path pointing .md metadata template}
}
\value{
tibble containing concatenated strings from each field
}
\description{
Read metadata md information into R
}
\examples{
path_to_md_template <- system.file("markdown/westerband_template.md", package = "correa")
metadata_ls <- read_metadata_md(path_to_md_template)
}
\keyword{internal}
