% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_workflow.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{suggest_workflow}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{corella}{\code{\link[corella]{suggest_workflow}}}
}}

