% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_str.R
\name{update_names}
\alias{update_names}
\title{Update Name Field}
\usage{
update_names(template, metadata_ls)
}
\arguments{
\item{template}{ALA template structure generated by \code{eml_template_ls}}

\item{metadata_ls}{Metadata list generated by \code{convert_md_tibble_to_list}}
}
\value{
updated nested list with updated field
}
\description{
Update Name Field
}
\examples{
eml <- eml_template_ls()
eml
update_names(eml,metadata_ls)
}
\keyword{internal}
