% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_data}
\alias{use_bd_data}
\title{Add \code{data} folder to a biodiversity data repository}
\usage{
use_bd_data(..., overwrite = FALSE)
}
\arguments{
\item{...}{Unquoted names of existing objects to save.}

\item{overwrite}{(logical) Should existing objects be overwritten? Defaults
to FALSE.}
}
\description{
This function places specified objects in the \code{data} folder as csv files.
Note that this is very different from \code{usethis::use_data} which uses \code{.rda}
format (and then only when \code{internal = FALSE}).
}
