% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_data}
\alias{use_bd_data}
\title{Add \code{data} folder to a Biodiversity Data Package}
\usage{
use_bd_data(..., overwrite = FALSE)
}
\arguments{
\item{...}{Unquoted names of existing objects to save.}

\item{overwrite}{(logical) Should existing objects be overwritten? Defaults
to FALSE.}
}
\description{
This function is largely synonymous with \code{usethis::use_data()}, but is
included here for completeness, and to enforce some defaults that affect
where and how data is stored.
}
