% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_description}
\alias{use_bd_description}
\title{Add \code{DESCRIPTION} to a biodiversity data repository}
\usage{
use_bd_description()
}
\description{
In a biodiversity data repository, the DESCRIPTION file is used to add
authorship and licencing information, which is then used by \code{build_dwca()}
to create a metadata statement (in conjunction with README.Rmd)
}
