% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_metadata}
\alias{use_bd_metadata}
\title{Add a metadata statement to a Biodiversity Data Package}
\usage{
use_bd_metadata()
}
\description{
Example metadata statements can be browsed at https://collections.ala.org.au
}
