% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_metadata}
\alias{use_bd_metadata}
\title{Add a metadata statement to a biodiversity data repository}
\usage{
use_bd_metadata()
}
\description{
Builds a file called \code{metadata.md} in the \code{inst} folder; this folder is
created if not already present. Partially populated using \code{DESCRIPTION}.
}
