% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_bd_readme}
\alias{use_bd_readme}
\alias{use_bd_readme_md}
\alias{use_bd_readme_rmd}
\title{Add \code{README} to a biodiversity data project}
\usage{
use_bd_readme_md()

use_bd_readme_rmd()
}
\description{
This function adds \code{galaxias}-specific \code{README} instead of the \code{usethis}
default.
}
