% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_data.R
\name{use_data}
\alias{use_data}
\title{Use standardised data in a Darwin Core Archive}
\usage{
use_data(..., type = c("occurrence", "event"), overwrite = FALSE)
}
\arguments{
\item{...}{(string) Unquoted name of \code{tibble}/\code{data.frame} to save.}

\item{type}{(string) Data type. Designates whether data are saved as
\code{occurrences.csv} or \code{events.csv}. \code{type} is determined by detecting
Event-type column names in data (only valid in an Event-type dataset).}

\item{overwrite}{(logical) By default, \code{use_data()} will not
overwrite existing files. If you really want to do so, set this to \code{TRUE}.}
}
\value{
Invisibly returns the location of the saved csv file.
}
\description{
Once data conform to Darwin Core Standard, \code{use_data()} makes it
easy to save data in the correct place for building a Darwin Core Archive
with \code{\link[=build_archive]{build_archive()}}.
}
\details{
This function saves data in the \code{data-publish} folder, which is the default
folder that \code{galaxias} uses to build a Darwin Core Archive using
\code{\link[=build_archive]{build_archive()}}.
}
\seealso{
\code{\link[=use_metadata]{use_metadata()}}
}
