% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_datetime.R
\name{use_datetime}
\alias{use_datetime}
\title{Add date and time fields to a \code{tibble}}
\usage{
use_datetime(
  .df,
  eventDate = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  eventTime = NULL,
  .keep = "unused",
  .messages = TRUE
)
}
\arguments{
\item{eventDate}{The date or date + time that the observation/event occurred.}

\item{year}{The year of the observation/event.}

\item{month}{The month of the observation/event.}

\item{day}{The day of the observation/event.}

\item{.keep}{Control which columns from .data are retained in the output.
Note that unlike \code{dplyr::mutate}, which defaults to \code{"all"} this defaults to
\code{"unused"}; i.e. only keeps Darwin Core fields, and not those fields used to
generate them.}

\item{df}{a \code{data.frame} or \code{tibble} that the column should be appended to.}

\item{time}{The time of the observation/event.}
}
\value{
A tibble with the requested fields added.
}
\description{
This function helps format standard date/time fields to a \code{tibble}.
}
\details{
In
practice this is no different from using \code{mutate()}, but gives some
informative errors, and serves as a useful lookup for how spatial fields are
represented in the Darwin Core Standard.

Example values are:
\itemize{
\item \code{eventDate} should be class \code{Date} or \code{POSITct}. We suggest using the
lubridate package to define define your date format using functions like
\code{ymd()}, \code{mdy}, \code{dmy()}, or if including date + time, \code{ymd_hms()},
\code{ymd_hm()}, or \code{ymd_h()}.
}
}
