% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dwc.R
\name{use_dwc}
\alias{use_dwc}
\alias{use_occurrenceID}
\alias{use_basisOfRecord}
\title{Use common Darwin Core fields}
\usage{
use_occurrenceID(.df, ...)

use_basisOfRecord(.df, value = c("humanObservation", "machineObservation"))
}
\description{
Shortcut methods for adding standard Darwin Core terms
In development, here for example purposes
}
