% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_metadata.R
\name{use_metadata}
\alias{use_metadata}
\title{Use a metadata statement in a Darwin Core Archive}
\usage{
use_metadata(
  source = "metadata.Rmd",
  destination = "data-publish/eml.xml",
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{source}{A metadata file in markdown (\code{.md}), Rmarkdown (\code{.Rmd}) or
Quarto markdown (\code{.qmd}). Defaults
to \code{metadata.md}, which is the same as is created by \code{\link[=use_metadata_template]{use_metadata_template()}}}

\item{destination}{A file name to save the resulting \code{.xml} file. Defaults to
\code{eml.xml}. Note that the file is saved to the \code{data-publish} directory,
unless changed using the \code{directory} argument of \code{\link[=galaxias_config]{galaxias_config()}}.}

\item{overwrite}{By default, \code{use_metadata()} will not
overwrite existing files. If you really want to do so, set this to \code{TRUE}.}

\item{quiet}{Whether to message about what is happening. Default is set to
\code{FALSE}.}
}
\value{
Does not return an object to the workspace; called for the side
effect of building a file in the \code{data-publish} directory.
}
\description{
A metadata statement lists the owner of the dataset, how it was collected,
and how it can be used (i.e. its' licence). This function reads and
converts metadata saved in markdown (.md), Rmarkdown (.Rmd) or Quarto (.qmd)
to xml, and saves it in the \code{destination} folder.

This function is a convenience wrapper function of \code{\link[delma:read_md]{delma::read_md()}} and
\code{\link[delma:write_eml]{delma::write_eml()}}.
}
\seealso{
\code{\link[=use_metadata_statement]{use_metadata_statement()}} to create a metadata statement template.
}
