% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_metadata.R
\name{use_metadata}
\alias{use_metadata}
\title{Write an example metadata statement}
\usage{
use_metadata(file, overwrite = FALSE)
}
\arguments{
\item{file}{(string) A filename to save the statement to. Defaults to
\code{"metadata.md"}.}

\item{overwrite}{(logical) Should any existing file with this name be
overwritten? Defaults to \code{FALSE}.}
}
\value{
Does not return an object to the workspace; called for the
side-effect of saving a markdown file to the specified location.
}
\description{
This function takes \code{metadata_example} and exports it to the specified file.
This is useful for creating a boilerplate metadata statement that you
can edit yourself.
}
