% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_schema.R
\name{use_schema}
\alias{use_schema}
\title{Create a \code{schema} for a Darwin Core Archive}
\usage{
use_schema(overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{overwrite}{By default, \code{use_schema()} will not overwrite existing
files. If you really want to do so, set this to \code{TRUE}.}

\item{quiet}{(logical) Should progress messages be suppressed? Default is
set to \code{FALSE}; i.e. messages are shown.}
}
\value{
Does not return an object to the workspace; called for the side
effect of building a schema file in the publication directory.
}
\description{
A schema is an xml document that maps the files and field names in a DwCA.
This map makes it easier to reconstruct one or more related datasets so that
information is matched correctly. It works by detecting column names on csv
files in a specified directory; these should all be Darwin Core terms for
this function to produce reliable results. This function assumes that the
publishing directory is named \code{"data-publish"}.
}
\details{
To be compliant with the Darwin Core Standard, the schema file \strong{must} be
called \code{meta.xml}, and this function enforces that.
}
\examples{
\dontrun{
use_schema()
}
}
