% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_archive.R
\name{validate_archive}
\alias{validate_archive}
\title{Validate a Darwin Core Archive via API}
\usage{
validate_archive(pkg = ".", file = NULL, provider = "GBIF")
}
\arguments{
\item{file}{Optionally specify a pre-built DwCA instead of a directory.}

\item{provider}{(string) The institution to be queried for validation
services. Currently only \code{"GBIF"} is supported.}

\item{path}{Path to a nominated archive.}
}
\value{
A report on the supplied archive.
}
\description{
Validate a Darwin Core Archive via API
}
