% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_archive.R, R/get_validator_report.R
\name{validate_archive}
\alias{validate_archive}
\alias{get_validator_report}
\title{Validate a Darwin Core Archive via API}
\usage{
validate_archive(x = "data", provider = "GBIF")

get_validator_report(key)
}
\arguments{
\item{x}{(string) Either a directory containing all the files to be stored in
the archive, or a filename (ending in .zip) of the specified archive.
Defaults to the \code{data} folder within the current working directory.}

\item{provider}{(string) The institution to be queried for validation
services. Currently only \code{"GBIF"} is supported.}

\item{key}{A unique identifier key from GBIF}
}
\value{
Invisibly returns a tibble to the workspace containing validation
results; but primarily called for the side-effect of generating a report in
the console.
}
\description{
Validation is a process of checking whether a specified archive is ready for
sharing and publication, according to the Darwin Core standard.
}
\seealso{
\code{check_archive()} which runs checks locally, rather than via API.
}
