% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_dwca.R
\name{validate_dwca}
\alias{validate_dwca}
\title{Validate a DwCA using the Atlas of Living Australia API}
\usage{
validate_dwca(path = ".", file = NULL)
}
\arguments{
\item{path}{path to a nominated Biodiversity Data Package}

\item{file}{optionally specify a pre-built DwCA instead of a directory}
}
\description{
Typically, most testing is done using \code{devtools::test}; sensible tests for
Darwin Core formats can be created using\code{add_bd_testthat()}. Prior to
publication, however, it is often useful to check suitability against the
host institutions criteria. This function uses the ALA to validate DwCAs.
}
