% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_md.R
\name{write_md}
\alias{write_md}
\title{Write a \code{tibble} to a markdown document}
\usage{
write_md(df, file)
}
\arguments{
\item{df}{A tibble in same format as \code{read_md()}}

\item{file}{either a character string naming a file, or a connection open for
writing.}
}
\description{
Export format is chosen based on the file suffix; GitHub
(.md), Rmarkdown (.Rmd) and Quarto (.qmd) are supported.
}
